Code data

Make structured code data set.

as.codedata(x, ...)

# S3 method for data.table
as.codedata(
  x,
  id = "id",
  code = "code",
  code_date = if (utils::hasName(x, "code_date")) "code_date" else NULL,
  period = NULL,
  .setkeys = TRUE,
  alnum = FALSE,
  ...
)

is.codedata(x)

Arguments

x
data table/frame with columns id (character), code, and optional code_date (Date).

...
arguments passed between methods

id, code
column names (from x) with case id:s (character) and codes of interest.

code_date
optional column name (from x) with relevant dates corresponding to each code. If there is a column named code_date this will be used by default.

period
optional date vector of length two with lower and upper limits for inclusion period

.setkeys
should output be indexed (Beneficial for large data sets.)

alnum
Should codes be cleaned from all non alphanumeric characters?

Value

data.table with columns: id (character), code and optionally code_date (Date) Possible additional columns from x are preserved.

Examples

x <-
  data.frame(
    id               = "1",
    code_date        = as.Date("2017-01-01"),
    code             = "a",
    stringsAsFactors = FALSE
  )
as.codedata(x)

#>    id  code_date code
#> 1:  1 2017-01-01    a

# Ignore codes for dates outside specified limits
y <-
  data.frame(
    id               = "2",
    code_date        = as.Date("3017-02-02"),
    code             = "b",
    stringsAsFactors = FALSE
  )
z <- rbind(x, y)
as.codedata(z, period = c(as.Date("1970-01-01"), Sys.Date()))

#>    id  code_date code
#> 1:  1 2017-01-01    a