Categorize cases based on external data and classification scheme

This is the main function of the package, which relies of a triad of objects: (1) data with unit id:s and possible dates of interest; (2) codedata for corresponding units and with optional dates of interest and; (3) a classification scheme (classcodes object; cc) with regular expressions to identify and categorize relevant codes. The function combines the three underlying steps performed by codify(), classify() and index(). Relevant arguments are passed to those functions by codify_args and cc_args.

categorize(x, ...)

# S3 method for data.frame
categorize(x, ...)

# S3 method for tbl_df
categorize(x, ...)

# S3 method for data.table
categorize(x, ..., codedata, id, code, codify_args = list())

# S3 method for codified
categorize(
  x,
  ...,
  cc,
  index = NULL,
  cc_args = list(),
  check.names = TRUE,
  .data_cols = NULL
)

Arguments

x
data set with mandatory character id column (identified by argument id = "<col_name>"), and optional Date of interest (identified by argument date = "<col_name>"). Alternatively, the output from codify()

...
arguments passed between methods

codedata
external code data with mandatory character id column (identified by id = "<col_name>"), code column (identified by argument code = "<col_name>") and optional Date column (identified by codify_args = list(code_date = "<col_name>")).

id
name of unique character id column found in both xand codedata. (where it must not be unique).

code
name of code column in codedata.

codify_args
Lists of named arguments passed to codify()

cc
classcodes object (or name of a default object from all_classcodes()).

index
Argument passed to index(). A character vector of names of columns with index weights from the corresponding classcodes object (as supplied by the ccargument). See attr(cc, "indices") for available options. Set to FALSE if no index should be calculated. If NULL, the default, all available indices (from attr(cc, "indices")) are provided.

cc_args
List with named arguments passed to set_classcodes()

check.names
Column names are based on cc$group, which might include spaces. Those names are changed to syntactically correct names by check.names = TRUE. Syntactically invalid, but grammatically correct names might be preferred for presentation of the data as achieved by check.names = FALSE. Alternatively, if categorize is called repeatedly, longer informative names might be created by cc_args = list(tech_names = TRUE).

.data_cols
used internally

Value

Object of the same class as x with additional logical columns indicating membership of groups identified by the classcodes object (the cc argument). Numeric indices are also included if requested by the index argument.

See also

Examples

# For some patient data (ex_people) and related hospital visit code data
# with ICD 10-codes (ex_icd10), add the Elixhauser comorbidity
# conditions based on all registered ICD10-codes
categorize(
   x            = ex_people,
   codedata     = ex_icd10,
   cc           = "elixhauser",
   id           = "name",
   code         = "icd10"
)

#> Classification based on: icd10

#> # A tibble: 100 x 40
#>    name        surgery    congestive.heart.f… cardiac.arrhythm… valvular.disease
#>    <chr>       <date>     <lgl>               <lgl>             <lgl>           
#>  1 Chen, Trev… 2021-01-19 FALSE               FALSE             FALSE           
#>  2 Graves, Ac… 2020-10-11 FALSE               FALSE             FALSE           
#>  3 Trujillo, … 2020-09-28 FALSE               FALSE             FALSE           
#>  4 Simpson, K… 2020-12-31 FALSE               FALSE             FALSE           
#>  5 Chin, Nels… 2020-12-14 FALSE               FALSE             FALSE           
#>  6 Le, Christ… 2020-07-18 FALSE               FALSE             FALSE           
#>  7 Kang, Xuan  2020-10-20 FALSE               FALSE             FALSE           
#>  8 Shuemaker,… 2020-07-19 FALSE               FALSE             FALSE           
#>  9 Boucher, T… 2020-12-25 FALSE               FALSE             FALSE           
#> 10 Le, Soraiya 2020-11-29 FALSE               FALSE             FALSE           
#> # … with 90 more rows, and 35 more variables:
#> #   pulmonary.circulation.disorder <lgl>, peripheral.vascular.disorder <lgl>,
#> #   hypertension.uncomplicated <lgl>, hypertension.complicated <lgl>,
#> #   paralysis <lgl>, other.neurological.disorders <lgl>,
#> #   chronic.pulmonary.disease <lgl>, diabetes.uncomplicated <lgl>,
#> #   diabetes.complicated <lgl>, hypothyroidism <lgl>, renal.failure <lgl>,
#> #   liver.disease <lgl>, peptic.ulcer.disease <lgl>, AIDS.HIV <lgl>,
#> #   lymphoma <lgl>, metastatic.cancer <lgl>, solid.tumor <lgl>,
#> #   rheumatoid.arthritis <lgl>, coagulopathy <lgl>, obesity <lgl>,
#> #   weight.loss <lgl>, fluid.electrolyte.disorders <lgl>,
#> #   blood.loss.anemia <lgl>, deficiency.anemia <lgl>, alcohol.abuse <lgl>,
#> #   drug.abuse <lgl>, psychoses <lgl>, depression <lgl>, sum_all <dbl>,
#> #   sum_all_ahrq <dbl>, walraven <dbl>, sid29 <dbl>, sid30 <dbl>,
#> #   ahrq_mort <dbl>, ahrq_readm <dbl>

# Add Charlson categories and two versions of a calculated index
# ("quan_original" and "quan_updated").
categorize(
   x            = ex_people,
   codedata     = ex_icd10,
   cc           = "charlson",
   id           = "name",
   code         = "icd10",
   index        = c("quan_original", "quan_updated")
)

#> Classification based on: icd10

#> # A tibble: 100 x 21
#>    name      surgery    myocardial.infar… congestive.heart.… peripheral.vascula…
#>    <chr>     <date>     <lgl>             <lgl>              <lgl>              
#>  1 Chen, Tr… 2021-01-19 FALSE             FALSE              FALSE              
#>  2 Graves, … 2020-10-11 FALSE             FALSE              FALSE              
#>  3 Trujillo… 2020-09-28 FALSE             FALSE              FALSE              
#>  4 Simpson,… 2020-12-31 FALSE             FALSE              FALSE              
#>  5 Chin, Ne… 2020-12-14 FALSE             FALSE              FALSE              
#>  6 Le, Chri… 2020-07-18 FALSE             FALSE              FALSE              
#>  7 Kang, Xu… 2020-10-20 FALSE             FALSE              FALSE              
#>  8 Shuemake… 2020-07-19 FALSE             FALSE              FALSE              
#>  9 Boucher,… 2020-12-25 FALSE             FALSE              TRUE               
#> 10 Le, Sora… 2020-11-29 FALSE             FALSE              FALSE              
#> # … with 90 more rows, and 16 more variables: cerebrovascular.disease <lgl>,
#> #   dementia <lgl>, chronic.pulmonary.disease <lgl>, rheumatic.disease <lgl>,
#> #   peptic.ulcer.disease <lgl>, mild.liver.disease <lgl>,
#> #   diabetes.without.complication <lgl>, hemiplegia.or.paraplegia <lgl>,
#> #   renal.disease <lgl>, diabetes.complication <lgl>, malignancy <lgl>,
#> #   moderate.or.severe.liver.disease <lgl>, metastatic.solid.tumor <lgl>,
#> #   AIDS.HIV <lgl>, quan_original <dbl>, quan_updated <dbl>

# Only include recent hospital visits within 30 days before surgery,
categorize(
   x            = ex_people,
   codedata     = ex_icd10,
   cc           = "charlson",
   id           = "name",
   code         = "icd10",
   index        = c("quan_original", "quan_updated"),
   codify_args  = list(
      date      = "surgery",
      days      = c(-30, -1),
      code_date = "admission"
   )
)

#> Classification based on: icd10

#> # A tibble: 100 x 21
#>    name      surgery    myocardial.infar… congestive.heart.… peripheral.vascula…
#>    <chr>     <date>     <lgl>             <lgl>              <lgl>              
#>  1 Chen, Tr… 2021-01-19 FALSE             FALSE              FALSE              
#>  2 Graves, … 2020-10-11 NA                NA                 NA                 
#>  3 Trujillo… 2020-09-28 NA                NA                 NA                 
#>  4 Simpson,… 2020-12-31 FALSE             FALSE              FALSE              
#>  5 Chin, Ne… 2020-12-14 FALSE             FALSE              FALSE              
#>  6 Le, Chri… 2020-07-18 FALSE             FALSE              FALSE              
#>  7 Kang, Xu… 2020-10-20 FALSE             FALSE              FALSE              
#>  8 Shuemake… 2020-07-19 FALSE             FALSE              FALSE              
#>  9 Boucher,… 2020-12-25 NA                NA                 NA                 
#> 10 Le, Sora… 2020-11-29 FALSE             FALSE              FALSE              
#> # … with 90 more rows, and 16 more variables: cerebrovascular.disease <lgl>,
#> #   dementia <lgl>, chronic.pulmonary.disease <lgl>, rheumatic.disease <lgl>,
#> #   peptic.ulcer.disease <lgl>, mild.liver.disease <lgl>,
#> #   diabetes.without.complication <lgl>, hemiplegia.or.paraplegia <lgl>,
#> #   renal.disease <lgl>, diabetes.complication <lgl>, malignancy <lgl>,
#> #   moderate.or.severe.liver.disease <lgl>, metastatic.solid.tumor <lgl>,
#> #   AIDS.HIV <lgl>, quan_original <dbl>, quan_updated <dbl>

# Multiple versions -------------------------------------------------------

# We can compare categorization by according to Quan et al. (2005); "icd10",
# and Armitage et al. (2010); "icd10_rcs" (see `?charlson`)
# Note the use of `tech_names = TRUE` to distinguish the column names from the
# two versions.

# We first specify some common settings ...
ind <- c("quan_original", "quan_updated")
cd  <- list(date = "surgery", days = c(-30, -1), code_date = "admission")

# ... we then categorize once with "icd10" as the default regular expression ...
categorize(
   x            = ex_people,
   codedata     = ex_icd10,
   cc           = "charlson",
   id           = "name",
   code         = "icd10",
   index        = ind,
   codify_args  = cd,
   cc_args      = list(tech_names = TRUE)
) %>%

# .. and once more with `regex = "icd10_rcs"`
categorize(
   codedata     = ex_icd10,
   cc           = "charlson",
   id           = "name",
   code         = "icd10",
   index        = ind,
   codify_args  = cd,
   cc_args      = list(regex = "icd10_rcs", tech_names = TRUE)
)

#> Classification based on: icd10

#> # A tibble: 100 x 37
#>    name    surgery    charlson_icd10_myo… charlson_icd10_co… charlson_icd10_per…
#>    <chr>   <date>     <lgl>               <lgl>              <lgl>              
#>  1 Chen, … 2021-01-19 FALSE               FALSE              FALSE              
#>  2 Graves… 2020-10-11 NA                  NA                 NA                 
#>  3 Trujil… 2020-09-28 NA                  NA                 NA                 
#>  4 Simpso… 2020-12-31 FALSE               FALSE              FALSE              
#>  5 Chin, … 2020-12-14 FALSE               FALSE              FALSE              
#>  6 Le, Ch… 2020-07-18 FALSE               FALSE              FALSE              
#>  7 Kang, … 2020-10-20 FALSE               FALSE              FALSE              
#>  8 Shuema… 2020-07-19 FALSE               FALSE              FALSE              
#>  9 Bouche… 2020-12-25 NA                  NA                 NA                 
#> 10 Le, So… 2020-11-29 FALSE               FALSE              FALSE              
#> # … with 90 more rows, and 32 more variables:
#> #   charlson_icd10_cerebrovascular_disease <lgl>,
#> #   charlson_icd10_dementia <lgl>,
#> #   charlson_icd10_chronic_pulmonary_disease <lgl>,
#> #   charlson_icd10_rheumatic_disease <lgl>,
#> #   charlson_icd10_peptic_ulcer_disease <lgl>,
#> #   charlson_icd10_mild_liver_disease <lgl>,
#> #   charlson_icd10_diabetes_without_complication <lgl>,
#> #   charlson_icd10_hemiplegia_or_paraplegia <lgl>,
#> #   charlson_icd10_renal_disease <lgl>,
#> #   charlson_icd10_diabetes_complication <lgl>,
#> #   charlson_icd10_malignancy <lgl>,
#> #   charlson_icd10_moderate_or_severe_liver_disease <lgl>,
#> #   charlson_icd10_metastatic_solid_tumor <lgl>, charlson_icd10_aids_hiv <lgl>,
#> #   charlson_icd10_index_quan_original <dbl>,
#> #   charlson_icd10_index_quan_updated <dbl>,
#> #   charlson_icd10_rcs_myocardial_infarction <lgl>,
#> #   charlson_icd10_rcs_congestive_heart_failure <lgl>,
#> #   charlson_icd10_rcs_peripheral_vascular_disease <lgl>,
#> #   charlson_icd10_rcs_cerebrovascular_disease <lgl>,
#> #   charlson_icd10_rcs_dementia <lgl>,
#> #   charlson_icd10_rcs_chronic_pulmonary_disease <lgl>,
#> #   charlson_icd10_rcs_rheumatic_disease <lgl>,
#> #   charlson_icd10_rcs_hemiplegia_or_paraplegia <lgl>,
#> #   charlson_icd10_rcs_renal_disease <lgl>,
#> #   charlson_icd10_rcs_diabetes_complication <lgl>,
#> #   charlson_icd10_rcs_malignancy <lgl>,
#> #   charlson_icd10_rcs_moderate_or_severe_liver_disease <lgl>,
#> #   charlson_icd10_rcs_metastatic_solid_tumor <lgl>,
#> #   charlson_icd10_rcs_aids_hiv <lgl>,
#> #   charlson_icd10_rcs_index_quan_original <dbl>,
#> #   charlson_icd10_rcs_index_quan_updated <dbl>

# column names ------------------------------------------------------------

# Default column names are based on row names from corresponding classcodes
# object but are modified to be syntactically correct.
default <-
   categorize(ex_people, codedata = ex_icd10, cc = "elixhauser",
              id = "name", code = "icd10")

#> Classification based on: icd10

# Set `check.names = FALSE` to retain original names:
original <-
  categorize(
    ex_people, codedata = ex_icd10, cc = "elixhauser",
    id = "name", code = "icd10",
    check.names = FALSE
   )

#> Classification based on: icd10

# Or use `tech_names = TRUE` for informative but long names (use case above)
tech <-
  categorize(ex_people, codedata = ex_icd10, cc = "elixhauser",
    id = "name", code = "icd10",
    cc_args = list(tech_names = TRUE)
  )

#> Classification based on: icd10

# Compare
tibble::tibble(names(default), names(original), names(tech))

#> # A tibble: 40 x 3
#>    `names(default)`        `names(original)`       `names(tech)`                
#>    <chr>                   <chr>                   <chr>                        
#>  1 name                    name                    name                         
#>  2 surgery                 surgery                 surgery                      
#>  3 congestive.heart.failu… congestive heart failu… elixhauser_icd10_congestive_…
#>  4 cardiac.arrhythmias     cardiac arrhythmias     elixhauser_icd10_cardiac_arr…
#>  5 valvular.disease        valvular disease        elixhauser_icd10_valvular_di…
#>  6 pulmonary.circulation.… pulmonary circulation … elixhauser_icd10_pulmonary_c…
#>  7 peripheral.vascular.di… peripheral vascular di… elixhauser_icd10_peripheral_…
#>  8 hypertension.uncomplic… hypertension uncomplic… elixhauser_icd10_hypertensio…
#>  9 hypertension.complicat… hypertension complicat… elixhauser_icd10_hypertensio…
#> 10 paralysis               paralysis               elixhauser_icd10_paralysis   
#> # … with 30 more rows