Code data

Code data

as.codedata(x, y = NULL, ..., .setkeys = TRUE, .copy = NA,
  npr = FALSE, nprdate = "utdatuma")

is.codedata(x)

Arguments

x
data frame with columns "id", "code_date" and "code" or object in format used by NPR (see below).

y
additional (optional) NPR data set if x contains data from NPR data. (see below).

...
by default, codes for future dates or dates before "1970-01-01" are ignored (with a warning). Specify to, from (as passed to filter_dates) to override.

.setkeys
should keys be set for the data.table object? To initially set the key for a large data set can be slow but it could be beneficial for later use of the object.

.copy
Should the object be copied internally by copy? NA (by default) means that objects smaller than 1 Gb are copied. If the size is larger, the argument must be set explicitly. Set TRUE to make copies regardless of object size. This is recomended if enough RAM is available. If set to FALSE, calculations might be carried out but the object will be changed by reference.

npr
Is data from the Swedish patient register? If so, the format might be recognised automatically (no guarantee).

nprdate
If x is recognised as a data set from the Swedish National Patient Register, which date variable should be reconised as code date for out patients? Could be either "indatuma" or "utdatuma". (Note that in patients only have one date.)

Value

as.codedata returns a data.table object with mandatory columns: "id" (individual id coerced to character) and "code". It has an optional column "code_date" (when the code was valid) if there was such a column in x or if the data is recognised from NPR. Possible additional columns from x are preserved if data is not recognised as coming from NPR.

is.codedata returns TRUE if these same conditions are met and FALSE otherwise. (Note that codedata is not a formal R class of its own!)

Data from NPR

Data from the Swedish Patient Register can be recognised as such if argument x contains columns named lpnr, indatuma/utdatuma AND (hdia, bdia1, ..., bdia15 OR op1, ..., op30).

Note that the name should be indatuma/utdatuma (with letter s suffixed). This is to distinguish the character date format %Y%m%d from the SPSS date format sometimes stored in indatum/utdatum. If the SPSS date is the only one available, the help function spss2date can be used for transformation, but this must be done before using as.codedata.

If the data set is of this format, it is transformed to the codedata format and returned as such. Also, NPR data often comes in the form of two separate data sets, one for inpatient and one for outpatient care. It is therefore possible to supply an additional data set by argument y. It is assumed that data from outpatient care has relevant dates stored as utdatuma and in-patient data has relevant dates named utdatuma. This is however not tested formally. If utdatuma is missing from the out-patient data, indatuma is used without warning.

Examples

x <- data.frame(
  id = 1,
  code_date = as.Date("2017-02-02"),
  code = "a"
)
as.codedata(x)

#>    id  code_date code
#> 1:  1 2017-02-02    a

# Drop dates outside specified limits
y <- data.frame(id = 2, code_date = as.Date("3017-02-02"), code = "b")
z <- rbind(x, y)
as.codedata(z)

#>    id  code_date code
#> 1:  1 2017-02-02    a