Example NPR data with random codes assigned to random people

Eictive ICD-10-codes in the format as delivered by the National Patient Register (NPR).

ex_pardata

Format

Data frames with 1,000 rows and 18 columns:

lpnr

Random names of people

date

random dates for inpatient registration

hdia

main diagnose given by ICD-10

bdia(1, ..., 14)

additional diagnoses for each visit

Details

Note that a real NPR data set would have "bdia" codes filled in order. The situation with no "bdia1" but with a "bdia2" is therefore not possible in a real data set.

See also