Visualise classification scheme in web browser

Classes are visualised by their regular expressions in a web browser. The visualisation does not give any details on group names, conditions or weights but might be useful both for understanding of a clasification scheme in use, and during the creation and debugging of such.

visualize(x, group = NULL, show = TRUE, ...)

visualise(x, group = NULL, show = TRUE, ...)

Arguments

x
classcodes specification as either a name or a classcodes object, or a classcodes object itself.

group
names (as character vector) of groups to visualise (all groups if NULL)

show
should a visualisation be shown in a web browser. Set to FALSE to just retrieve a URL for later use.

...
Arguments passed on to get_classcodes

regex

name of column with regular expressions to use for classification, either with or without prefix regex_

Value

URL to website with visualisation (invisible)

See also

Examples

if (FALSE) {

 # How is depression classified according to Elixhauser?
 visualize("elixhauser", "depression")

 # Compare the two diabetes groups according to Charlson
 visualize("charlson",
   c("diabetes without complication", "diabetes complication"))

 # Is this different from the "Royal College of Surgeons classification?
 # Yes, there is only one group for diabetes
 visualize("charlson",
   c("diabetes without complication", "diabetes complication"),
   regex = "rcs"
 )

 # Show all groups from Charlson
 visualize("charlson")

 # Get URL for later visualisation
 (visualize("tha.fracture.ae", show = FALSE))

}