Classcodes for adverse events after knee and hip arthroplasty based on ICD-10 and KVA codes

classcodes objects to use with classify

knee.ae_icd10

hip.ae_icd10

hip.ae_kva

knee.ae_kva

hip.fracture.ae_icd10

Format

Data frame with 3 columns:

group

Different types of adverse events (see reference section)

regex

regular expressions identifying ICD-10 or KVA codes for each group

condition

special conditions are used, see below.

Details

Group names are prefixed by two letters as given by the references. Two groups (DB and DM) are split into two due to different conditions.

Hip fractures

AE codes for hip fractures are based on codes for elective cases but with some additional codes for DM 1 (N300, N308, N309 and N390).

Conditions

Special conditions apply to all categories. Those require non-standard modifications of the classcodes data prior to categorization.

hbdia1_hdia

TRUE if the code was given as any type of diagnose during hospital visit for index operation, or as main diagnose for later visits, otherwise FALSE

late_hdia

TRUE if the code was given as main diagnose at a later visit after the index operation, otherwise FALSE

post_op

TRUE if the code was given at a later visit after the index operation, otherwise FALSE

References

Knee (p. 83): http://www.myknee.se/pdf/SVK-2016_1.1.pdf. Hip (p. 149): https://registercentrum.blob.core.windows.net/shpr/r/-rsrapport-2017-S1xKMzsAwX.pdf

See also