Categorize cases based on external data and classification scheme

This is the main function of the package, which relies of a triad of objects: (1) data with unit id:s and possible dates of interest; (2) codedata for corresponding units and with optional dates of interest and; (3) a classification scheme (classcodes object; cc) with regular expressions to identify and categorize relevant codes.

categorize(x, ...)

# S3 method for data.frame
categorize(x, ...)

# S3 method for tbl_df
categorize(x, ...)

# S3 method for data.table
categorize(x, ..., codedata, id, code, codify_args = list())

# S3 method for codified
categorize(
  x,
  ...,
  cc,
  index = NULL,
  sort = TRUE,
  cc_args = list(),
  .data_cols = NULL
)

Arguments

x
data.frame or data.table::data.table with mandatory id column (identified by argument id), and optional column with date of interest (identified by argument date if days != NULL). Alternatively, the output from codify()

...
arguments passed between methods

codedata
external code data (see prepare_codedata())

id
name of code/id columns (in codified).

code
name of code/id columns (in codified).

codify_args
List of named arguments passed to codify()

cc
classcodes object (or name of such object).

index
A character vector of index values to calculate (passed to argument by in function index().

Set to FALSE if no index should be calculated.

If NULL, the default, all available indices (from attr(cc, "indices")) are provided. A message lists the indices so that you can check they're correct; suppress the message by supplying index explicitly.

sort
logical. Should output be sorted by the 'id' column? (This could affect computational speed for large data sets. Data is sorted by 'id' internally. It is therefore faster to keep the output sorted this way, but this might be inconvenient if the original order was important.)

cc_args
List with named arguments passed to set_classcodes()

.data_cols
used internally

Value

Object of class data.table made from data combined with logical columns indicating membership of categories identified by the classcodes object (the cc argument). Indices are also included if specified by the 'index' argument.

Details

The function combines the three underlying steps performed by codify(), classify() and index(). Relevant arguments are passed to those functions by codify_args and cc_args.

See also

Examples

# Add Elixhauser based on all registered ICD10-codes
categorize(ex_people, ex_icd10, "elixhauser", id = "name")

#> Error in categorize.data.table(as.data.table(x), ...): argument "codedata" is missing, with no default

# Add Charlson categories and two versions of a calculated index.
# Only include recent hospital visits within 30 days before surgery,
# and use technical variable names to clearly identify the new columns.
categorize(ex_people, ex_icd10, "charlson",
  id = "name",
  index = c("quan_original", "quan_updated"),
  codify_args = list(date = "event", days = c(-30, -1)),
  cc_args = list(tech_names = TRUE)
)

#> Error in categorize.data.table(as.data.table(x), ...): argument "codedata" is missing, with no default