Categorize cases based on external data and classification scheme

This is the main function of the package, which relies of a triad of objects: (1) data with unit id:s and possible dates of interest; (2) codedata for corresponding units and with optional dates of interest and; (3) a classification scheme (classcodes object; cc) with regular expressions to identify and categorize relevant codes.

categorize(x, ...)

# S3 method for data.frame
categorize(x, ...)

# S3 method for tbl_df
categorize(x, ...)

# S3 method for data.table
categorize(x, ..., codedata, id, code, codify_args = list())

# S3 method for codified
categorize(
  x,
  ...,
  cc,
  index = NULL,
  sort = TRUE,
  cc_args = list(),
  check.names = TRUE,
  .data_cols = NULL
)

Arguments

x
data set with mandatory id column (identified by argument id), and optional column with date of interest (identified by argument date if days != NULL). Alternatively, the output from codify()

...
arguments passed between methods

codedata
external code data

id
name of code/id columns (in codified).

code
name of code/id columns (in codified).

cc
classcodes object (or name of such object).

index
A character vector of index values to calculate (passed to index(). Set to FALSE if no index should be calculated. If NULL, the default, all available indices (from attr(cc, "indices")) are provided. A message lists the indices so that you can check they're correct; suppress the message by supplying index explicitly.

sort
logical. Should output be sorted by the 'id' column? (This could affect computational speed for large data sets. Data is sorted by 'id' internally. It is therefore faster to keep the output sorted this way, but this might be inconvenient if the original order was important.)

cc_args, codify_args
List of named arguments passed to set_classcodes() and codify()

check.names
Column names are based on cc$group, which might include spaces. Those names are changed to syntactically correct names by check.names = TRUE. Syntactically invalid, but grammatically correct names might be preferred for presentation of the data as achieved by check.names = FALSE. Alternatively, if categorize is called repeatedly, longer informative names might be created by cc_args = list(tech_names = TRUE).

.data_cols
used internally

Value

Object of the same class as x with additional logical columns indicating membership of categories identified by the classcodes object (the cc argument). Indices are also included if specified by the 'index' argument.

Details

The function combines the three underlying steps performed by codify(), classify() and index(). Relevant arguments are passed to those functions by codify_args and cc_args. index

See also

Examples

# Add Elixhauser based on all registered ICD10-codes
categorize(ex_people, codedata = ex_icd10, cc = "elixhauser",
  id = "name", code = "icd10")

#> Classification based on: icd10

#> # A tibble: 100 x 40
#>    name  surgery    `congestive hea~ `cardiac arrhyt~ `valvular disea~
#>    <chr> <date>     <lgl>            <lgl>            <lgl>           
#>  1 Arch~ 2020-05-25 FALSE            FALSE            FALSE           
#>  2 Awtr~ 2020-09-20 FALSE            FALSE            FALSE           
#>  3 Bamm~ 2020-03-25 FALSE            FALSE            FALSE           
#>  4 Bank~ 2020-06-18 FALSE            FALSE            FALSE           
#>  5 Bega~ 2020-06-19 FALSE            FALSE            FALSE           
#>  6 Beni~ 2020-02-24 FALSE            FALSE            FALSE           
#>  7 Bhat~ 2020-10-12 FALSE            FALSE            FALSE           
#>  8 Bouc~ 2020-09-04 FALSE            FALSE            FALSE           
#>  9 Carn~ 2020-09-11 FALSE            FALSE            TRUE            
#> 10 Chap~ 2020-05-18 FALSE            FALSE            FALSE           
#> # ... with 90 more rows, and 35 more variables: `pulmonary circulation
#> #   disorder` <lgl>, `peripheral vascular disorder` <lgl>, `hypertension
#> #   uncomplicated` <lgl>, `hypertension complicated` <lgl>, paralysis <lgl>,
#> #   `other neurological disorders` <lgl>, `chronic pulmonary disease` <lgl>,
#> #   `diabetes uncomplicated` <lgl>, `diabetes complicated` <lgl>,
#> #   hypothyroidism <lgl>, `renal failure` <lgl>, `liver disease` <lgl>, `peptic
#> #   ulcer disease` <lgl>, `AIDS/HIV` <lgl>, lymphoma <lgl>, `metastatic
#> #   cancer` <lgl>, `solid tumor` <lgl>, `rheumatoid arthritis` <lgl>,
#> #   coagulopathy <lgl>, obesity <lgl>, `weight loss` <lgl>, `fluid electrolyte
#> #   disorders` <lgl>, `blood loss anemia` <lgl>, `deficiency anemia` <lgl>,
#> #   `alcohol abuse` <lgl>, `drug abuse` <lgl>, psychoses <lgl>,
#> #   depression <lgl>, sum_all <dbl>, sum_all_ahrq <dbl>, walraven <dbl>,
#> #   sid29 <dbl>, sid30 <dbl>, ahrq_mort <dbl>, ahrq_readm <dbl>

# Add Charlson categories and two versions of a calculated index.
# Only include recent hospital visits within 30 days before surgery,
# and use technical variable names to clearly identify the new columns.
categorize(ex_people, codedata = ex_icd10, cc = "charlson",
  id = "name", code = "icd10",
  index = c("quan_original", "quan_updated"),
  codify_args =
    list(date = "surgery", days = c(-30, -1), code_date = "admission"),
  cc_args = list(tech_names = TRUE)
)

#> Classification based on: icd10

#> # A tibble: 100 x 21
#>    name  surgery    charlson_icd10_~ charlson_icd10_~ charlson_icd10_~
#>    <chr> <date>     <lgl>            <lgl>            <lgl>           
#>  1 Arch~ 2020-05-25 NA               NA               NA              
#>  2 Awtr~ 2020-09-20 FALSE            FALSE            FALSE           
#>  3 Bamm~ 2020-03-25 FALSE            FALSE            FALSE           
#>  4 Bank~ 2020-06-18 FALSE            FALSE            FALSE           
#>  5 Bega~ 2020-06-19 FALSE            FALSE            FALSE           
#>  6 Beni~ 2020-02-24 NA               NA               NA              
#>  7 Bhat~ 2020-10-12 NA               NA               NA              
#>  8 Bouc~ 2020-09-04 NA               NA               NA              
#>  9 Carn~ 2020-09-11 NA               NA               NA              
#> 10 Chap~ 2020-05-18 NA               NA               NA              
#> # ... with 90 more rows, and 16 more variables:
#> #   charlson_icd10_cerebrovascular_disease <lgl>,
#> #   charlson_icd10_dementia <lgl>,
#> #   charlson_icd10_chronic_pulmonary_disease <lgl>,
#> #   charlson_icd10_rheumatic_disease <lgl>,
#> #   charlson_icd10_peptic_ulcer_disease <lgl>,
#> #   charlson_icd10_mild_liver_disease <lgl>,
#> #   charlson_icd10_diabetes_without_complication <lgl>,
#> #   charlson_icd10_hemiplegia_or_paraplegia <lgl>,
#> #   charlson_icd10_renal_disease <lgl>,
#> #   charlson_icd10_diabetes_complication <lgl>,
#> #   charlson_icd10_malignancy <lgl>,
#> #   charlson_icd10_moderate_or_severe_liver_disease <lgl>,
#> #   charlson_icd10_metastatic_solid_tumor <lgl>, charlson_icd10_aids_hiv <lgl>,
#> #   charlson_icd10_index_quan_original <dbl>,
#> #   charlson_icd10_index_quan_updated <dbl>