Classcodes methods

"Classcodes" are classifying schemes based on regular expression stored in data frames. These are essential to the package and constitute the third part of the triad of case data, code data and classification scheme.

as.classcodes(x, hierarchy = attr(x, "hierarchy"), .name = NULL)

is.classcodes(x)

Arguments

x
data frame with properties as described in the details section

hierarchy
list of pairwise group names to appear as superior and subordinate classes. To be used for indexing when the subordinate class is redundant.

.name
used internally for name dispatch

Value

Object of class "classcodes" (data frame) with attribute code specifying the coding used (for example "icd10", or "ATC"). Could be NULL for unknown or arbitrary coding.

Details

A classcodes object is a data frame with columns:

group:

unique and non missing class name

regex:

regular expression defining class membership (see regex for details). Occurrences of non unique regular expressions will lead to the same class having multiple names. This is accepted but will raise a warning.

description (optional):

description of each category

condition (optional):

a class might have conditions additional to what is expressed by regex. If so, these should be specified as quoted expressions that can be evaluated within the data frame used by classify

weights (optional):

weights for each class used for index calculation. Could be more than one and could have arbitrary names.

hierarchy

pairwise hierarchys for wich only the superior class should be used for indexing

Note that classes does not have to be disjunct.

The package have several default classcodes included listed on its index page `help(package = "coder")`

See also

Examples

as.classcodes(coder::elixhauser)

#> # A tibble: 31 x 13
#>    group regex_icd10 regex_icd10_sho~ regex_icd9cm regex_icd9cm_ah~
#>  * <chr> <chr>       <chr>            <chr>        <chr>           
#>  1 cong~ I(099|1(10~ I(09|1[13]|25|4~ 39891|4(0(2~ 39891|4(0(2[019~
#>  2 card~ I(44[1-3]|~ I(4[457-9])|R00~ 42(6(1[013]~ NA              
#>  3 valv~ A520|I(0[5~ A52|I(0[5-9]|3[~ 0932|39([4-~ 0932|39([4-6][0~
#>  4 pulm~ I(2([67]|8~ I2[678]          41(6|79)     41(6|79)        
#>  5 peri~ I7([01]|3[~ I7[01389]|K55|Z~ 44(0|1[2479~ 44([0-2]|3[1-9]~
#>  6 hype~ I10         I10              401[19]      401[19]|6420    
#>  7 hype~ I1[1-35]    I1[1-35]         40([24][19]~ 40(10|[2-5])|64~
#>  8 para~ G(041|114|~ G(04|11|8[0-3])  34(2[019]|[~ 34[2-4]|438[2-5]
#>  9 othe~ G(1[0-3]|2~ G(1[0-3]|2[0125~ 3(3(19|20|3~ 3(3([0145]|20|3~
#> 10 chro~ (I27[89])|~ I27|(J([46][0-7~ 49(([01]|2[~ 49|50([0-5]|6[0~
#> # ... with 21 more rows, and 8 more variables: regex_icd9cm_enhanced <chr>,
#> #   index_sum_all <dbl>, index_sum_all_ahrq <dbl>, index_walraven <dbl>,
#> #   index_sid29 <dbl>, index_sid30 <dbl>, index_ahrq_mort <dbl>,
#> #   index_ahrq_readm <dbl>