Classcodes methods

"Classcodes" are classifying schemes based on regular expression stored in data frames. These are essential to the package and constitute the third part of the triad of case data, code data and classification scheme.

as.classcodes(x)

is.classcodes(x)

Arguments

x
data frame with properties as described in the details section

Value

Object of class "classcodes" (data frame) with attribute code specifying the coding used (for example "icd10", or "ATC"). Could be NULL for unknown or arbitrary coding.

Details

A classcodes object is a data frame with columns:

group:

unique and non missing class name

regex:

regular expression defining class membership (see regex for details). Occurrences of non unique regular expressions will lead to the same class having multiple names. This is accepted but will raise a warning.

description (optional):

description of each category

condition (optional):

a class might have conditions additional to what is expressed by regex. If so, these should be specified as quoted expressions that can be evaluated within the data frame used by classify

weights (optional):

weights for each class used for index calculation. Could be more than one and could have arbitrary names.

Note that classes does not have to be disjunct.

The package have several default classcodes included listed on its index page `help(package = "coder")`

See also

Examples

as.classcodes(coder::elixhauser)

#> # A tibble: 31 x 10
#>    group regex regex_short sum_all sum_all_ahrq walraven sid29 sid30 ahrq_mort
#>  * <chr> <chr> <chr>         <dbl>        <dbl>    <dbl> <dbl> <dbl>     <dbl>
#>  1 cong… ^(I(… ^(I(09|1[1…       1            1        7     9     9         9
#>  2 card… ^(I(… ^(I(4[457-…       1            0        5     8     0         0
#>  3 valv… ^(A5… ^(A52|I(0[…       1            1       -1     0     0         0
#>  4 pulm… ^I(2… ^I2[678]          1            1        4     5     5         6
#>  5 peri… ^(I7… ^(I7[01389…       1            1        2     4     4         3
#>  6 hype… ^I10  ^I10              1            1        0    -2    -1        -1
#>  7 hype… ^I1[… ^I1[1-35]         1            1        0    -2    -1        -1
#>  8 para… ^G(0… ^G(04|11|8…       1            1        7     4     3         5
#>  9 othe… ^(G(… ^(G(1[0-3]…       1            1        6     5     4         5
#> 10 chro… ^((I… ^(I27|(J([…       1            1        3     3     3         3
#> # … with 21 more rows, and 1 more variable: ahrq_readm <dbl>