Classcodes methods

"Classcodes" are classifying schemes based on regular expression stored in data frames. These are essential to the package and constitute the third part of the triad of case data, code data and classification scheme.

as.classcodes(x, coding = NULL)

is.classcodes(x)

Arguments

x
data frame with properties as described in the details section

coding
Coding used for the classification (character vector of length one). Used by summary.classcodes.

Value

Object of class "classcodes" (data frame) with attribute code specifying the coding used (for example "icd10", or "ATC"). Could be NULL for unknown or arbitrary coding.

Details

A classcodes object is a data frame with columns:

group:

unique and non missing class name

regex:

regular expression defining class membership (see regex for details). Occurrences of non unique regular expressions will lead to the same class having multiple names. This is accepted but will raise a warning.

condition (optional):

a class might have conditions additional to what is expressed by regex. If so, these should be specified as quoted expressions that can be evaluated within the data frame used by classify

weights (optional):

weights for each class used for index calculation. Could be more than one and could have arbitrary names.

Note that classes does not have to be disjunct.

The package have several default classcodes included listed on its index page `help(package = "coder")`

See also

Examples

as.classcodes(coder::elix_icd10)

#> # A tibble: 31 x 2
#>    group                    regex                                              
#>  * <chr>                    <chr>                                              
#>  1 congestive heart failure ^(I099|I1(10|3[02])|I255|I4(2[05-9]|3)|I50|P290)   
#>  2 cardiac arrhythmias      ^(I44[1-3]|I456|I459|I4[7-9]|R00[018]|T821|Z[49]50)
#>  3 valvular disease         ^(A520|I0[5-8]|I09[18]|I3[4-9]|Q23[0-3]|Z95[2-4])  
#>  4 pulmonary circulation d~ ^(I2([67]|8[089]))                                 
#>  5 peripheral vascular dis~ ^(I7([01]|3[189]|71|9[02])|K55[189]|Z95[89])       
#>  6 hypertension uncomplica~ ^(I10[[:digit:]])                                  
#>  7 hypertension complicated ^(I1[1-35])                                        
#>  8 paralysis                ^(G041|G114|G8(0[12]|[12]|3[0-49]))                
#>  9 other neurological diso~ ^(G1[0-3]|G2[012]|G25[45]|G31[289]|G3[25-7]|G4[01]~
#> 10 chronic pulmonary disea~ ^(I27[89]|J4[0-7]|J6([0-7]|84)|J70[13])            
#> # ... with 21 more rows