Classify cases based on classcodes

Classify cases based on classcodes

classify(codified, cc, ..., cc_args = list())

# S3 method for default
classify(codified, cc, ..., cc_args = list())

# S3 method for codified
classify(codified, ...)

# S3 method for data.frame
classify(codified, ...)

# S3 method for data.table
classify(codified, cc, ..., id, code, cc_args = list())

Arguments

codified
output from codify()

cc
classcodes object or name of such

...
arguments passed between methods

cc_args
List with named arguments passed to set_classcodes()

code, id
name of code/id columns (in codified).

Value

Boolean matrix with one row for each element/row of codified and columns for each class with corresponding class names (according to the classcodes object).

See also

Examples

# Classify individual ICD10-codes by Elixhauser
classify(c("C80", "I20", "unvalid_code"), "elixhauser")

#> Classification based on: regex_icd10

#>              congestive heart failure cardiac arrhythmias valvular disease
#> C80                             FALSE               FALSE            FALSE
#> I20                             FALSE               FALSE            FALSE
#> unvalid_code                    FALSE               FALSE            FALSE
#>              pulmonary circulation disorder peripheral vascular disorder
#> C80                                   FALSE                        FALSE
#> I20                                   FALSE                        FALSE
#> unvalid_code                          FALSE                        FALSE
#>              hypertension uncomplicated hypertension complicated paralysis
#> C80                               FALSE                    FALSE     FALSE
#> I20                               FALSE                    FALSE     FALSE
#> unvalid_code                      FALSE                    FALSE     FALSE
#>              other neurological disorders chronic pulmonary disease
#> C80                                 FALSE                     FALSE
#> I20                                 FALSE                     FALSE
#> unvalid_code                        FALSE                     FALSE
#>              diabetes uncomplicated diabetes complicated hypothyroidism
#> C80                           FALSE                FALSE          FALSE
#> I20                           FALSE                FALSE          FALSE
#> unvalid_code                  FALSE                FALSE          FALSE
#>              renal failure liver disease peptic ulcer disease AIDS/HIV lymphoma
#> C80                  FALSE         FALSE                FALSE    FALSE    FALSE
#> I20                  FALSE         FALSE                FALSE    FALSE    FALSE
#> unvalid_code         FALSE         FALSE                FALSE    FALSE    FALSE
#>              metastatic cancer solid tumor rheumatoid arthritis coagulopathy
#> C80                       TRUE       FALSE                FALSE        FALSE
#> I20                      FALSE       FALSE                FALSE        FALSE
#> unvalid_code             FALSE       FALSE                FALSE        FALSE
#>              obesity weight loss fluid electrolyte disorders blood loss anemia
#> C80            FALSE       FALSE                       FALSE             FALSE
#> I20            FALSE       FALSE                       FALSE             FALSE
#> unvalid_code   FALSE       FALSE                       FALSE             FALSE
#>              deficiency anemia alcohol abuse drug abuse psychoses depression
#> C80                      FALSE         FALSE      FALSE     FALSE      FALSE
#> I20                      FALSE         FALSE      FALSE     FALSE      FALSE
#> unvalid_code             FALSE         FALSE      FALSE     FALSE      FALSE
#> attr(,"classcodes")
#> [1] "elixhauser"
#> attr(,"id")
#> [1] "id"
#> attr(,"class")
#> [1] "classified" "matrix"    

# Classify patients by Charlson for comorbidities during
# one year before surgery
x <- codify(ex_people, ex_icd10, id = "name",
  date = "event", days = c(-365, 0))

#> Error in utils::hasName(codedata, code): argument "code" is missing, with no default

y <- classify(x, "charlson")

#> Error in classify(x, "charlson"): object 'x' not found

# Use tha RCS classification instead and use technical column names
y <- classify(x, "charlson",
  cc_args = list(regex = "icd10_rcs", tech_names = TRUE))

#> Error in classify(x, "charlson", cc_args = list(regex = "icd10_rcs", tech_names = TRUE)): object 'x' not found

# It is possible to convert the output of classify to a data frame with
# id column instead of row names
as.data.frame(y)

#> Error in as.data.frame(y): object 'y' not found