Make codebook for classcodes object

Make codebook for classcodes object

codebook(object, coding, ..., file = NULL)

Arguments

object
classcodes object

coding
either a vector with codes from the original classification, or a name (character vector of length one) keyvalue object from package "decoder" (for example "icd10cm" or "atc")

...
Arguments passed on to summary.classcodes

cc_args

List of named arguments passed to set_classcodes()

file
name of Excel file for data export

Value

List of data frames describing relationship between groups and individual codes

See also

Examples

# All codes from ICD-10-CM used by Elixhauser
codebook(elixhauser, "icd10cm")

#> Classification based on: icd10

#> Coodebooks are prefarably exported to Excel using the `file` argument! Use `summary.classcodes()` or `visualize()` for interactive summaries!

#> Preview of README sheet specifying columns in following sheets:
#> 
#> # A tibble: 6 x 3
#>   tab       column      description                                       
#>   <chr>     <chr>       <chr>                                             
#> 1 summary   group       Group to categorize by                            
#> 2 summary   n           Number of identified codes contained in this group
#> 3 summary   codes       Comma separated list of individual codes in group 
#> 4 all_codes code        Individual code                                   
#> 5 all_codes description Description of individual code                    
#> 6 all_codes group       Group to categorize by                            
#> 
#> 
#> Preview of summary sheet:
#> 
#> # A tibble: 31 x 3
#>    group                   n codes                                              
#>    <chr>               <int> <chr>                                              
#>  1 AIDS/HIV                1 B20                                                
#>  2 alcohol abuse         131 E52, F1010, F1011, F10120, F10121, F10129, F1014, ~
#>  3 blood loss anemia       1 D500                                               
#>  4 cardiac arrhythmias    73 I441, I442, I4430, I4439, I456, I459, I470, I471, ~
#>  5 chronic pulmonary ~    69 I27812, I27822, I27832, I27892, I2792, J40, J410, ~
#>  6 coagulopathy           27 D65, D66, D67, D680, D681, D682, D68311, D68312, D~
#>  7 congestive heart f~    36 I099, I1101, I1301, I1321, I255, I420, I425, I4261~
#>  8 deficiency anemia      17 D508, D509, D510, D511, D512, D513, D518, D519, D5~
#>  9 depression             31 F3130, F3131, F3132, F314, F3152, F320, F321, F322~
#> 10 diabetes complicat~   243 E1021, E1022, E1029, E10311, E10319, E103211, E103~
#> # ... with 21 more rows
#> 
#> 
#> Preview of all_codes:
#> 
#> # A tibble: 3,127 x 3
#>    code  description                                   group           
#>    <chr> <chr>                                         <chr>           
#>  1 A5200 Cardiovascular syphilis, unspecified          valvular disease
#>  2 A5201 Syphilitic aneurysm of aorta                  valvular disease
#>  3 A5202 Syphilitic aortitis                           valvular disease
#>  4 A5203 Syphilitic endocarditis                       valvular disease
#>  5 A5204 Syphilitic cerebral arteritis                 valvular disease
#>  6 A5205 Other cerebrovascular syphilis                valvular disease
#>  7 A5206 Other syphilitic heart involvement            valvular disease
#>  8 A5209 Other cardiovascular syphilis                 valvular disease
#>  9 B180  Chronic viral hepatitis B with delta-agent    liver disease   
#> 10 B181  Chronic viral hepatitis B without delta-agent liver disease   
#> # ... with 3,117 more rows

# All codes from ICD-9-CM Disease part used by Elixhauser enhanced version
codebook(elixhauser, "icd9cmd",
  cc_args = list(regex = "icd9cm_enhanced")
)

#> Coodebooks are prefarably exported to Excel using the `file` argument! Use `summary.classcodes()` or `visualize()` for interactive summaries!

#> Preview of README sheet specifying columns in following sheets:
#> 
#> # A tibble: 6 x 3
#>   tab       column      description                                       
#>   <chr>     <chr>       <chr>                                             
#> 1 summary   group       Group to categorize by                            
#> 2 summary   n           Number of identified codes contained in this group
#> 3 summary   codes       Comma separated list of individual codes in group 
#> 4 all_codes code        Individual code                                   
#> 5 all_codes description Description of individual code                    
#> 6 all_codes group       Group to categorize by                            
#> 
#> 
#> Preview of summary sheet:
#> 
#> # A tibble: 31 x 3
#>    group                   n codes                                              
#>    <chr>               <int> <chr>                                              
#>  1 AIDS/HIV                1 042                                                
#>  2 alcohol abuse          36 2652, 2911, 2912, 2913, 2915, 29181, 29182, 29189,~
#>  3 blood loss anemia       1 2800                                               
#>  4 cardiac arrhythmias    42 4260, 42610, 42611, 42612, 42613, 4262, 4263, 4264~
#>  5 chronic pulmonary ~    16 41682, 41692, 490, 500, 501, 502, 503, 504, 505, 5~
#>  6 coagulopathy           20 2860, 2861, 2862, 2863, 2864, 28652, 28653, 28659,~
#>  7 congestive heart f~    32 39891, 402011, 402111, 402911, 404011, 404031, 404~
#>  8 deficiency anemia      10 2801, 2808, 2809, 2810, 2811, 2812, 2813, 2814, 28~
#>  9 depression             38 29620, 29621, 29622, 29623, 29624, 29625, 29626, 2~
#> 10 diabetes complicat~    24 25040, 25041, 25042, 25043, 25050, 25051, 25052, 2~
#> # ... with 21 more rows
#> 
#> 
#> Preview of all_codes:
#> 
#> # A tibble: 1,251 x 3
#>    code  description              group                       
#>    <chr> <chr>                    <chr>                       
#>  1 042   Human immuno virus dis   AIDS/HIV                    
#>  2 07022 Hpt B chrn coma wo dlta  liver disease               
#>  3 07023 Hpt B chrn coma w dlta   liver disease               
#>  4 07032 Hpt B chrn wo cm wo dlta liver disease               
#>  5 07033 Hpt B chrn wo cm w dlta  liver disease               
#>  6 07044 Chrnc hpt C w hepat Coma liver disease               
#>  7 07054 Chrnc hpt C wo hpat coma liver disease               
#>  8 0706  Viral hepat NOS w coma   liver disease               
#>  9 0709  Viral hepat NOS w/o coma liver disease               
#> 10 0930  Aortic aneurysm, syphil  peripheral vascular disorder
#> # ... with 1,241 more rows