codebook(s) for classcodes object

summary.classcodes() and visualize.classcodes() are used to summarize/visualize classcodes in R. A codebook, on the other hand, is an exported summary saved in an Excel spreadsheet to use in collaboration with non R-users. Several codebooks might be combined into a single Excel document with several sheets (one for each codebook).

codebook(object, coding, ..., file = NULL)

# S3 method for codebook
print(x, ...)

codebooks(..., file = NULL)

Arguments

object
classcodes object

coding
either a vector with codes from the original classification, or a name (character vector of length one) keyvalue object from package "decoder" (for example "icd10cm" or "atc")

...
Additional arguments for each function:

codebook(): arguments passed to summary.classcodes()

codebooks(): multiple named outputs from codebook()

print.codebook(): arguments passed to tibble:::print.tbl()

file
name/path to Excel file for data export

x
output from codebook()

Value

Functions are primarily called for their side effects (exporting data to Excel or printing to screen). In addition:

codebook()returns list of data frames describing relationship between groups and individual codes

codebooks() returns a concatenated list with output from codebook(). Only one 'README' object is kept however and renamed as such.

print.codebook()returns x (invisible)

See also

Examples

# codebook() --------------------------------------------------------------
if (FALSE) {
# Export codebook (to temporary file) with all codes identified by the
# Elixhauser comorbidity classification based on ICD-10-CM
codebook(elixhauser, "icd10cm", file = tempfile("codebook", fileext = ".xlsx"))

# All codes from ICD-9-CM Disease part used by Elixhauser enhanced version
codebook(elixhauser, "icd9cmd",
  cc_args = list(regex = "icd9cm_enhanced",
  file = tempfile("codebook", fileext = ".xlsx"))
)

# The codebook returns a list with three objects.
# Access a dictionary table with translates of each code to text:
codebook(charlson, "icd10cm")$all_codes
}

# print.codebook() --------------------------------------------------------

# If argument `file` is unspecified, a preview of each sheet of the codebook is
# printed to the screen
(cb <- codebook(charlson, "icd10cm"))

#> Classification based on: icd10

#> Coodebooks are prefarably exported to Excel using the `file` argument! Use `summary.classcodes()` or `visualize()` for interactive summaries!

#> Preview of README sheet specifying columns in following sheets:
#> 
#> # A tibble: 6 x 3
#>   tab       column      description                                       
#>   <chr>     <chr>       <chr>                                             
#> 1 summary   group       Group to categorize by                            
#> 2 summary   n           Number of identified codes contained in this group
#> 3 summary   codes       Comma separated list of individual codes in group 
#> 4 all_codes code        Individual code                                   
#> 5 all_codes description Description of individual code                    
#> 6 all_codes group       Group to categorize by                            
#> 
#> 
#> Preview of summary sheet:
#> 
#> # A tibble: 17 x 3
#>    group                      n codes                                           
#>    <chr>                  <int> <chr>                                           
#>  1 AIDS/HIV                   1 B20                                             
#>  2 cerebrovascular disea…   430 G450, G451, G452, G453, G454, G458, G459, G460,…
#>  3 chronic pulmonary dis…    69 I2781, I2782, I2783, I2789, I279, J40, J410, J4…
#>  4 congestive heart fail…    36 I099, I110, I130, I132, I255, I420, I425, I426,…
#>  5 dementia                  11 F0150, F0151, F0280, F0281, F0390, F0391, G300,…
#>  6 diabetes complication    204 E1021, E1022, E1029, E10311, E10319, E103211, E…
#>  7 diabetes without comp…    52 E1010, E1011, E10610, E10618, E10620, E10621, E…
#>  8 hemiplegia or paraple…    45 G041, G114, G801, G802, G8100, G8101, G8102, G8…
#>  9 malignancy               961 C000, C001, C002, C003, C004, C005, C006, C008,…
#> 10 metastatic solid tumor    47 C770, C771, C772, C773, C774, C775, C778, C779,…
#> 11 mild liver disease        38 B180, B181, B182, B188, B189, K700, K7010, K701…
#> 12 moderate or severe li…    14 I8500, I8501, I864, K7040, K7041, K7110, K7111,…
#> 13 myocardial infarction     18 I2101, I2102, I2109, I2111, I2119, I2121, I2129…
#> 14 peptic ulcer disease      36 K250, K251, K252, K253, K254, K255, K256, K257,…
#> 15 peripheral vascular d…   274 I700, I701, I70201, I70202, I70203, I70208, I70…
#> 16 renal disease             28 I120, I1310, I1311, N032, N033, N034, N035, N03…
#> 17 rheumatic disease        348 M0500, M05011, M05012, M05019, M05021, M05022, …
#> 
#> 
#> Preview of all_codes:
#> 
#> # A tibble: 2,612 x 3
#>    code  description                                     group             
#>    <chr> <chr>                                           <chr>             
#>  1 B180  Chronic viral hepatitis B with delta-agent      mild liver disease
#>  2 B181  Chronic viral hepatitis B without delta-agent   mild liver disease
#>  3 B182  Chronic viral hepatitis C                       mild liver disease
#>  4 B188  Other chronic viral hepatitis                   mild liver disease
#>  5 B189  Chronic viral hepatitis, unspecified            mild liver disease
#>  6 B20   Human immunodeficiency virus [HIV] disease      AIDS/HIV          
#>  7 C000  Malignant neoplasm of external upper lip        malignancy        
#>  8 C001  Malignant neoplasm of external lower lip        malignancy        
#>  9 C002  Malignant neoplasm of external lip, unspecified malignancy        
#> 10 C003  Malignant neoplasm of upper lip, inner aspect   malignancy        
#> # … with 2,602 more rows

# The preview can be modified by arguments to the print-method
print(cb, n = 20)

#> Coodebooks are prefarably exported to Excel using the `file` argument! Use `summary.classcodes()` or `visualize()` for interactive summaries!

#> Preview of README sheet specifying columns in following sheets:
#> 
#> # A tibble: 6 x 3
#>   tab       column      description                                       
#>   <chr>     <chr>       <chr>                                             
#> 1 summary   group       Group to categorize by                            
#> 2 summary   n           Number of identified codes contained in this group
#> 3 summary   codes       Comma separated list of individual codes in group 
#> 4 all_codes code        Individual code                                   
#> 5 all_codes description Description of individual code                    
#> 6 all_codes group       Group to categorize by                            
#> 
#> 
#> Preview of summary sheet:
#> 
#> # A tibble: 17 x 3
#>    group                      n codes                                           
#>    <chr>                  <int> <chr>                                           
#>  1 AIDS/HIV                   1 B20                                             
#>  2 cerebrovascular disea…   430 G450, G451, G452, G453, G454, G458, G459, G460,…
#>  3 chronic pulmonary dis…    69 I2781, I2782, I2783, I2789, I279, J40, J410, J4…
#>  4 congestive heart fail…    36 I099, I110, I130, I132, I255, I420, I425, I426,…
#>  5 dementia                  11 F0150, F0151, F0280, F0281, F0390, F0391, G300,…
#>  6 diabetes complication    204 E1021, E1022, E1029, E10311, E10319, E103211, E…
#>  7 diabetes without comp…    52 E1010, E1011, E10610, E10618, E10620, E10621, E…
#>  8 hemiplegia or paraple…    45 G041, G114, G801, G802, G8100, G8101, G8102, G8…
#>  9 malignancy               961 C000, C001, C002, C003, C004, C005, C006, C008,…
#> 10 metastatic solid tumor    47 C770, C771, C772, C773, C774, C775, C778, C779,…
#> 11 mild liver disease        38 B180, B181, B182, B188, B189, K700, K7010, K701…
#> 12 moderate or severe li…    14 I8500, I8501, I864, K7040, K7041, K7110, K7111,…
#> 13 myocardial infarction     18 I2101, I2102, I2109, I2111, I2119, I2121, I2129…
#> 14 peptic ulcer disease      36 K250, K251, K252, K253, K254, K255, K256, K257,…
#> 15 peripheral vascular d…   274 I700, I701, I70201, I70202, I70203, I70208, I70…
#> 16 renal disease             28 I120, I1310, I1311, N032, N033, N034, N035, N03…
#> 17 rheumatic disease        348 M0500, M05011, M05012, M05019, M05021, M05022, …
#> 
#> 
#> Preview of all_codes:
#> 
#> # A tibble: 2,612 x 3
#>    code  description                                              group         
#>    <chr> <chr>                                                    <chr>         
#>  1 B180  Chronic viral hepatitis B with delta-agent               mild liver di…
#>  2 B181  Chronic viral hepatitis B without delta-agent            mild liver di…
#>  3 B182  Chronic viral hepatitis C                                mild liver di…
#>  4 B188  Other chronic viral hepatitis                            mild liver di…
#>  5 B189  Chronic viral hepatitis, unspecified                     mild liver di…
#>  6 B20   Human immunodeficiency virus [HIV] disease               AIDS/HIV      
#>  7 C000  Malignant neoplasm of external upper lip                 malignancy    
#>  8 C001  Malignant neoplasm of external lower lip                 malignancy    
#>  9 C002  Malignant neoplasm of external lip, unspecified          malignancy    
#> 10 C003  Malignant neoplasm of upper lip, inner aspect            malignancy    
#> 11 C004  Malignant neoplasm of lower lip, inner aspect            malignancy    
#> 12 C005  Malignant neoplasm of lip, unspecified, inner aspect     malignancy    
#> 13 C006  Malignant neoplasm of commissure of lip, unspecified     malignancy    
#> 14 C008  Malignant neoplasm of overlapping sites of lip           malignancy    
#> 15 C009  Malignant neoplasm of lip, unspecified                   malignancy    
#> 16 C01   Malignant neoplasm of base of tongue                     malignancy    
#> 17 C020  Malignant neoplasm of dorsal surface of tongue           malignancy    
#> 18 C021  Malignant neoplasm of border of tongue                   malignancy    
#> 19 C022  Malignant neoplasm of ventral surface of tongue          malignancy    
#> 20 C023  Malignant neoplasm of anterior two-thirds of tongue, pa… malignancy    
#> # … with 2,592 more rows

# codebooks() -------------------------------------------------------------

# Combine codebooks based on different versions of the regular expressions
# and export to a single (temporary) Excel file
c1 <- codebook(elixhauser, "icd10cm")

#> Classification based on: icd10

c2 <- codebook(elixhauser, "icd9cmd",
  cc_args = list(regex = "icd9cm_enhanced")
  )

codebooks(
  elix_icd10 = c1, elix_icd9cm = c2,
  file = tempfile("codebooks", fileext = ".xlsx")
)

#> codebooks saved as /var/folders/dd/z0l8zzy51jv48ywmfrv8yqt40000gp/T//Rtmp7uM2Ye/codebooks15bcd41e9b219.xlsx