Codify elements (within period)

Enhance case data with code data, possibly limited to relevant period.

codify(
  data,
  codedata,
  ...,
  id,
  code,
  date = NULL,
  code_date = NULL,
  days = NULL
)

# S3 method for data.frame
codify(data, ..., id, date = NULL, days = NULL)

# S3 method for data.table
codify(
  data,
  codedata,
  ...,
  id,
  code,
  date = NULL,
  code_date = NULL,
  days = NULL,
  alnum = FALSE,
  .copy = NA
)

# S3 method for codified
print(x, ..., n = 10)

Arguments

data
a table with columns id (character), code, and optionally date (Date)

codedata
additional data with columns including case id (character), code and an optional date (Date) for each code. An optional column condition might distinguish codes/dates with certain characteristics (see example).

...
arguments passed between methods

id, code, date, code_date
column names with case id (character from data and codedata), code (from data) and optional date (Date from data) and code_date (Date from codedata).

days
numeric vector of length two with lower and upper bound for range of relevant days relative to date. See "Relevant period".

alnum
Should codes be cleaned from all non alphanumeric characters?

.copy
Should the object be copied internally by data.table::copy()? NA (by default) means that objects smaller than 1 GB are copied. If the size is larger, the argument must be set explicitly. Set TRUE to make copies regardless of object size. This is recommended if enough RAM is available. If set to FALSE, calculations might be carried out but the object will be changed by reference.

x
object (potentially of large size)

n
number of rows to preview as tibble. The output is technically a data.table::data.table, which might be an unusual format to look at. Use n = NULL to print the object as is.

Value

Object of class codified (inheriting from data.table::data.table). Essentially data with additional columns: code, code_date: left joined from codedata or NA if no match within period. in_period: Boolean indicator if the case had at least one code within the specified period.

The output has one row for each combination of "id" from data and "code" from codedata. Rows from data might be repeated accordingly.

Details

See the example section for a warning on using the .copy argument!

Relevant period

Some examples for argument days:

c(-365, -1): window of one year prior to the date column of data. Useful for patient co-morbidity.

c(1, 30): window of 30 days after date. Useful for adverse events after a surgical procedure.

c(-Inf, Inf): no limitation on non missing dates.

NULL: no time limitation at all.

See also

Examples

# Codify all patients from `ex_people` with their ICD-10 codes from `ex_icd10`
x <- codify(ex_people, ex_icd10, id = "name", code = "icd10")
x

#> 
#> The printed data is of class: codified, data.table, data.frame.
#> It has 700 row(s).
#> It is here previewed as a tibble
#> Use `print(x, n = NULL)` to print as is (or use `n` to specify the number of rows to preview)!
#> 
#> 
#> # A tibble: 10 x 6
#>    name                admission  icd10 hdia  in_period surgery   
#>    <chr>               <date>     <chr> <lgl> <lgl>     <date>    
#>  1 Archer, Leon Hunter 2020-07-28 B469  FALSE TRUE      2020-05-25
#>  2 Archer, Leon Hunter 2019-12-06 E012  FALSE TRUE      2020-05-25
#>  3 Archer, Leon Hunter 2020-08-14 R900  FALSE TRUE      2020-05-25
#>  4 Archer, Leon Hunter 2020-02-11 V7413 FALSE TRUE      2020-05-25
#>  5 Archer, Leon Hunter 2019-12-28 V8698 FALSE TRUE      2020-05-25
#>  6 Archer, Leon Hunter 2020-01-18 X3403 FALSE TRUE      2020-05-25
#>  7 Archer, Leon Hunter 2020-01-17 X4128 FALSE TRUE      2020-05-25
#>  8 Archer, Leon Hunter 2020-02-03 Z752  FALSE TRUE      2020-05-25
#>  9 Awtrey, Antonio     2020-09-25 N608  FALSE TRUE      2020-09-20
#> 10 Awtrey, Antonio     2020-07-28 W0341 FALSE TRUE      2020-09-20

# Only consider codes if recorded at hospital admissions within one year prior
# to surgery
codify(
  ex_people,
  ex_icd10,
  id        = "name",
  code      = "icd10",
  date      = "surgery",
  code_date = "admission",
  days      = c(-365, 0)   # admission during one year before surgery
)

#> 
#> The printed data is of class: codified, data.table, data.frame.
#> It has 378 row(s).
#> It is here previewed as a tibble
#> Use `print(x, n = NULL)` to print as is (or use `n` to specify the number of rows to preview)!
#> 
#> 
#> # A tibble: 10 x 6
#>    name                surgery    admission  icd10 hdia  in_period
#>    <chr>               <date>     <date>     <chr> <lgl> <lgl>    
#>  1 Archer, Leon Hunter 2020-05-25 2019-12-06 E012  FALSE TRUE     
#>  2 Archer, Leon Hunter 2020-05-25 2019-12-28 V8698 FALSE TRUE     
#>  3 Archer, Leon Hunter 2020-05-25 2020-01-17 X4128 FALSE TRUE     
#>  4 Archer, Leon Hunter 2020-05-25 2020-01-18 X3403 FALSE TRUE     
#>  5 Archer, Leon Hunter 2020-05-25 2020-02-03 Z752  FALSE TRUE     
#>  6 Archer, Leon Hunter 2020-05-25 2020-02-11 V7413 FALSE TRUE     
#>  7 Awtrey, Antonio     2020-09-20 2020-01-14 X3322 FALSE TRUE     
#>  8 Awtrey, Antonio     2020-09-20 2020-04-05 Y1614 FALSE TRUE     
#>  9 Awtrey, Antonio     2020-09-20 2020-04-08 X7564 FALSE TRUE     
#> 10 Awtrey, Antonio     2020-09-20 2020-05-26 X6542 FALSE TRUE     

# Only consider codes if recorded after surgery
codify(
  ex_people,
  ex_icd10,
  id        = "name",
  code      = "icd10",
  date      = "surgery",
  code_date = "admission",
  days      = c(1, Inf)     # admission any time after surgery
)

#> 
#> The printed data is of class: codified, data.table, data.frame.
#> It has 355 row(s).
#> It is here previewed as a tibble
#> Use `print(x, n = NULL)` to print as is (or use `n` to specify the number of rows to preview)!
#> 
#> 
#> # A tibble: 10 x 6
#>    name                surgery    admission  icd10 hdia  in_period
#>    <chr>               <date>     <date>     <chr> <lgl> <lgl>    
#>  1 Archer, Leon Hunter 2020-05-25 2020-07-28 B469  FALSE TRUE     
#>  2 Archer, Leon Hunter 2020-05-25 2020-08-14 R900  FALSE TRUE     
#>  3 Awtrey, Antonio     2020-09-20 2020-09-25 N608  FALSE TRUE     
#>  4 Bammesberger, Jozi  2020-03-25 2020-04-28 V4931 FALSE TRUE     
#>  5 Bammesberger, Jozi  2020-03-25 2020-05-12 V4960 FALSE TRUE     
#>  6 Bammesberger, Jozi  2020-03-25 2020-07-02 X6414 FALSE TRUE     
#>  7 Bammesberger, Jozi  2020-03-25 2020-07-21 Y1513 FALSE TRUE     
#>  8 Bammesberger, Jozi  2020-03-25 2020-10-09 P293A FALSE TRUE     
#>  9 Banks, Silbret      2020-06-18 2020-07-06 D229D FALSE TRUE     
#> 10 Banks, Silbret      2020-06-18 2020-07-26 V7452 TRUE  TRUE     

# Dirty code data ---------------------------------------------------------

# Assume that codes contain unwanted "dirty" characters
# Those could for example be a dot used by ICD-10 (i.e. X12.3 instead of X123)
dirt <- c(strsplit(c("½!#¤%&/()=?`,.-_"), split = ""), recursive = TRUE)
rdirt <- function(x) sample(x, nrow(ex_icd10), replace = TRUE)
sub <- function(i) substr(ex_icd10$icd10, i, i)
ex_icd10$icd10 <-
  paste0(
    rdirt(dirt), sub(1),
    rdirt(dirt), sub(2),
    rdirt(dirt), sub(3),
    rdirt(dirt), sub(4),
    rdirt(dirt), sub(5)
  )
head(ex_icd10)

#> # A tibble: 6 x 4
#>   name                 admission  icd10      hdia 
#>   <chr>                <date>     <chr>      <lgl>
#> 1 Tran, Kenneth        2020-04-12 ,S/1-3!4&A FALSE
#> 2 Tran, Kenneth        2020-09-26 /W)3#3,1-9 FALSE
#> 3 Tran, Kenneth        2020-09-05 `Y(0&2%6?2 TRUE 
#> 4 Tran, Kenneth        2020-07-29 -X`0&4&8_8 FALSE
#> 5 Sommerville, Dominic 2020-09-17 %V.8_1-0`4 FALSE
#> 6 Sommerville, Dominic 2020-04-28 &B#8#5,3_  FALSE

# Use `alnum = TRUE` to ignore non alphanumeric characters
codify(ex_people, ex_icd10, id = "name", code = "icd10", alnum = TRUE)

#> 
#> The printed data is of class: codified, data.table, data.frame.
#> It has 700 row(s).
#> It is here previewed as a tibble
#> Use `print(x, n = NULL)` to print as is (or use `n` to specify the number of rows to preview)!
#> 
#> 
#> # A tibble: 10 x 6
#>    name                admission  icd10 hdia  in_period surgery   
#>    <chr>               <date>     <chr> <lgl> <lgl>     <date>    
#>  1 Archer, Leon Hunter 2020-07-28 B469  FALSE TRUE      2020-05-25
#>  2 Archer, Leon Hunter 2019-12-06 E012  FALSE TRUE      2020-05-25
#>  3 Archer, Leon Hunter 2020-08-14 R900  FALSE TRUE      2020-05-25
#>  4 Archer, Leon Hunter 2019-12-28 V8698 FALSE TRUE      2020-05-25
#>  5 Archer, Leon Hunter 2020-01-18 X3403 FALSE TRUE      2020-05-25
#>  6 Archer, Leon Hunter 2020-02-03 Z752  FALSE TRUE      2020-05-25
#>  7 Archer, Leon Hunter 2020-01-17 X4128 FALSE TRUE      2020-05-25
#>  8 Archer, Leon Hunter 2020-02-11 V7413 FALSE TRUE      2020-05-25
#>  9 Awtrey, Antonio     2020-05-26 X6542 FALSE TRUE      2020-09-20
#> 10 Awtrey, Antonio     2020-08-28 X4078 FALSE TRUE      2020-09-20

# Big data ----------------------------------------------------------------

# If `data` or `codedata` are large compared to available
# Random Access Memory (RAM) it might not be possible to make internal copies
# of those objects. Setting `.copy = FALSE` might help to overcome such problems

# If no copies are made internally, however, the input objects (if data tables)
# would change in the global environment
x2 <- data.table::as.data.table(ex_icd10)
head(x2) # Look at the "icd10" column (with dirty data)

#>                    name  admission      icd10  hdia
#> 1:        Tran, Kenneth 2020-04-12 ,S/1-3!4&A FALSE
#> 2:        Tran, Kenneth 2020-09-26 /W)3#3,1-9 FALSE
#> 3:        Tran, Kenneth 2020-09-05 `Y(0&2%6?2  TRUE
#> 4:        Tran, Kenneth 2020-07-29 -X`0&4&8_8 FALSE
#> 5: Sommerville, Dominic 2020-09-17 %V.8_1-0`4 FALSE
#> 6: Sommerville, Dominic 2020-04-28  &B#8#5,3_ FALSE

# Use `alnum = TRUE` combined with `.copy = FALSE`
codify(ex_people, x2, id = "name", code = "icd10", alnum = TRUE, .copy = FALSE)

#> 
#> The printed data is of class: codified, data.table, data.frame.
#> It has 700 row(s).
#> It is here previewed as a tibble
#> Use `print(x, n = NULL)` to print as is (or use `n` to specify the number of rows to preview)!
#> 
#> 
#> # A tibble: 10 x 6
#>    name                admission  icd10 hdia  in_period surgery   
#>    <chr>               <date>     <chr> <lgl> <lgl>     <date>    
#>  1 Archer, Leon Hunter 2020-08-14 R900  FALSE TRUE      2020-05-25
#>  2 Archer, Leon Hunter 2020-01-18 X3403 FALSE TRUE      2020-05-25
#>  3 Archer, Leon Hunter 2020-01-17 X4128 FALSE TRUE      2020-05-25
#>  4 Archer, Leon Hunter 2020-02-03 Z752  FALSE TRUE      2020-05-25
#>  5 Archer, Leon Hunter 2019-12-06 E012  FALSE TRUE      2020-05-25
#>  6 Archer, Leon Hunter 2020-02-11 V7413 FALSE TRUE      2020-05-25
#>  7 Archer, Leon Hunter 2019-12-28 V8698 FALSE TRUE      2020-05-25
#>  8 Archer, Leon Hunter 2020-07-28 B469  FALSE TRUE      2020-05-25
#>  9 Awtrey, Antonio     2020-05-26 X6542 FALSE TRUE      2020-09-20
#> 10 Awtrey, Antonio     2020-08-28 X4078 FALSE TRUE      2020-09-20

# Even though no explicit assignment was specified
# (neither for the output of codify(), nor to explicitly alter `x2`,
# the `x2` object has changed (look at the "icd10" column!):
head(x2)

#>                    name  admission icd10  hdia
#> 1:        Tran, Kenneth 2020-04-12 S134A FALSE
#> 2:        Tran, Kenneth 2020-09-26 W3319 FALSE
#> 3:        Tran, Kenneth 2020-09-05 Y0262  TRUE
#> 4:        Tran, Kenneth 2020-07-29 X0488 FALSE
#> 5: Sommerville, Dominic 2020-09-17 V8104 FALSE
#> 6: Sommerville, Dominic 2020-04-28  B853 FALSE

# Hence, the `.copy` argument should only be used if necessary
# and if so, with caution!

# print.codify() ----------------------------------------------------------

x # Preview first 10 rows as a tibble

#> 
#> The printed data is of class: codified, data.table, data.frame.
#> It has 700 row(s).
#> It is here previewed as a tibble
#> Use `print(x, n = NULL)` to print as is (or use `n` to specify the number of rows to preview)!
#> 
#> 
#> # A tibble: 10 x 6
#>    name                admission  icd10 hdia  in_period surgery   
#>    <chr>               <date>     <chr> <lgl> <lgl>     <date>    
#>  1 Archer, Leon Hunter 2020-07-28 B469  FALSE TRUE      2020-05-25
#>  2 Archer, Leon Hunter 2019-12-06 E012  FALSE TRUE      2020-05-25
#>  3 Archer, Leon Hunter 2020-08-14 R900  FALSE TRUE      2020-05-25
#>  4 Archer, Leon Hunter 2020-02-11 V7413 FALSE TRUE      2020-05-25
#>  5 Archer, Leon Hunter 2019-12-28 V8698 FALSE TRUE      2020-05-25
#>  6 Archer, Leon Hunter 2020-01-18 X3403 FALSE TRUE      2020-05-25
#>  7 Archer, Leon Hunter 2020-01-17 X4128 FALSE TRUE      2020-05-25
#>  8 Archer, Leon Hunter 2020-02-03 Z752  FALSE TRUE      2020-05-25
#>  9 Awtrey, Antonio     2020-09-25 N608  FALSE TRUE      2020-09-20
#> 10 Awtrey, Antonio     2020-07-28 W0341 FALSE TRUE      2020-09-20

print(x, n = 20) # Preview first 20 rows as a tibble

#> 
#> The printed data is of class: codified, data.table, data.frame.
#> It has 700 row(s).
#> It is here previewed as a tibble
#> Use `print(x, n = NULL)` to print as is (or use `n` to specify the number of rows to preview)!
#> 
#> 
#> # A tibble: 20 x 6
#>    name                admission  icd10 hdia  in_period surgery   
#>    <chr>               <date>     <chr> <lgl> <lgl>     <date>    
#>  1 Archer, Leon Hunter 2020-07-28 B469  FALSE TRUE      2020-05-25
#>  2 Archer, Leon Hunter 2019-12-06 E012  FALSE TRUE      2020-05-25
#>  3 Archer, Leon Hunter 2020-08-14 R900  FALSE TRUE      2020-05-25
#>  4 Archer, Leon Hunter 2020-02-11 V7413 FALSE TRUE      2020-05-25
#>  5 Archer, Leon Hunter 2019-12-28 V8698 FALSE TRUE      2020-05-25
#>  6 Archer, Leon Hunter 2020-01-18 X3403 FALSE TRUE      2020-05-25
#>  7 Archer, Leon Hunter 2020-01-17 X4128 FALSE TRUE      2020-05-25
#>  8 Archer, Leon Hunter 2020-02-03 Z752  FALSE TRUE      2020-05-25
#>  9 Awtrey, Antonio     2020-09-25 N608  FALSE TRUE      2020-09-20
#> 10 Awtrey, Antonio     2020-07-28 W0341 FALSE TRUE      2020-09-20
#> 11 Awtrey, Antonio     2020-01-14 X3322 FALSE TRUE      2020-09-20
#> 12 Awtrey, Antonio     2020-08-28 X4078 FALSE TRUE      2020-09-20
#> 13 Awtrey, Antonio     2020-05-26 X6542 FALSE TRUE      2020-09-20
#> 14 Awtrey, Antonio     2020-04-08 X7564 FALSE TRUE      2020-09-20
#> 15 Awtrey, Antonio     2020-07-22 Y0492 FALSE TRUE      2020-09-20
#> 16 Awtrey, Antonio     2020-04-05 Y1614 FALSE TRUE      2020-09-20
#> 17 Bammesberger, Jozi  2020-10-09 P293A FALSE TRUE      2020-03-25
#> 18 Bammesberger, Jozi  2019-11-21 V1051 FALSE TRUE      2020-03-25
#> 19 Bammesberger, Jozi  2020-01-28 V1392 FALSE TRUE      2020-03-25
#> 20 Bammesberger, Jozi  2020-04-28 V4931 FALSE TRUE      2020-03-25

print(x, n = NULL) # Print as data.table (ignoring the 'classified' class)

#>                     name  admission icd10  hdia in_period    surgery
#>   1: Archer, Leon Hunter 2020-07-28  B469 FALSE      TRUE 2020-05-25
#>   2: Archer, Leon Hunter 2019-12-06  E012 FALSE      TRUE 2020-05-25
#>   3: Archer, Leon Hunter 2020-08-14  R900 FALSE      TRUE 2020-05-25
#>   4: Archer, Leon Hunter 2020-02-11 V7413 FALSE      TRUE 2020-05-25
#>   5: Archer, Leon Hunter 2019-12-28 V8698 FALSE      TRUE 2020-05-25
#>  ---                                                                
#> 696:     el-Sani, Jawaad 2020-08-02  Y989  TRUE      TRUE 2020-08-30
#> 697:   el-Shahin, Hasnaa 2020-04-20 E711B FALSE      TRUE 2020-03-14
#> 698:   el-Shahin, Hasnaa 2019-10-28 X5144 FALSE      TRUE 2020-03-14
#> 699:   el-Shahin, Hasnaa 2020-01-28 Y586W FALSE      TRUE 2020-03-14
#> 700:   el-Shahin, Hasnaa 2020-02-14  Z888 FALSE      TRUE 2020-03-14