Codify elements (within period)

Enhance case data with code data, possibly limited to relevant period.

codify(data, codedata, id = "id", date = NULL, days = NULL, .copy = NA, ...)

Arguments

data
data.frame with at least case id, and optional date of interest

codedata
output from as.codedata.

id, date
column names (from data) with case id (character) and date of interest (optional).

days
numeric vector of length two with lower and upper bound for range of relevant days relative to date. See "Relevant period".

.copy
Should the object be copied internally by copy? NA (by default) means that objects smaller than 1 GB are copied. If the size is larger, the argument must be set explicitly. Set TRUE to make copies regardless of object size. This is recommended if enough RAM is available. If set to FALSE, calculations might be carried out but the object will be changed by reference.

...
Arguments passed on to as.codedata

alnum

Should codes be cleaned from all non alphanumeric characters?

Value

data (coerced to data.table) with additional columns: code, code_date: left joined from codedata or NA if no match within period. in_period: Boolean indicator if the case had at least one code within the specified period.

The output has one row for each combination of "id" from data and "code" from codedata. Rows from data might be repeated accordingly.

Relevant period

Some examples for argument days:

c(-365, -1): window of one year prior to the date column of data. Useful for patient co-morbidity.

c(1, 30): window of 30 days after date. Useful for adverse events after a surgical procedure.

c(-Inf, Inf): no limitation on non missing dates.

NULL: no time limitation at all.

See also

Examples

codify(ex_people, ex_icd10, id = "name", date = "event", days = c(-365, 0))

#>                    name       date  code_date  code  hdia in_period
#>   1: Alexander, Bethany 2019-03-07 2019-01-25  S925 FALSE      TRUE
#>   2: Alexander, Bethany 2019-03-07 2019-03-01  E278 FALSE      TRUE
#>   3:    Arguello, Tommy 2019-12-03 2018-12-14  H059 FALSE      TRUE
#>   4:    Arguello, Tommy 2019-12-03 2019-05-15  A231 FALSE      TRUE
#>   5:    Arguello, Tommy 2019-12-03 2019-08-21 Z032A FALSE      TRUE
#>  ---                                                               
#> 365:   el-Zaki, Huwaida 2019-12-21 2019-02-06  I712 FALSE      TRUE
#> 366:   el-Zaki, Huwaida 2019-12-21 2019-02-06  T336 FALSE      TRUE
#> 367:   el-Zaki, Huwaida 2019-12-21 2019-04-23  B958 FALSE      TRUE
#> 368:   el-Zaki, Huwaida 2019-12-21 2019-06-10 S3201 FALSE      TRUE
#> 369:   el-Zaki, Huwaida 2019-12-21 2019-11-17  C759 FALSE      TRUE