Codify elements (within period)

Enhance case data with code data, possibly limited to relevant period.

codify(
  data,
  codedata,
  ...,
  id,
  code,
  date = NULL,
  code_date = NULL,
  days = NULL
)

# S3 method for data.frame
codify(data, ..., id, date = NULL, days = NULL)

# S3 method for data.table
codify(
  data,
  codedata,
  ...,
  id,
  code,
  date = NULL,
  code_date = NULL,
  days = NULL,
  alnum = FALSE,
  .copy = NA
)

# S3 method for codified
print(x, ..., n = 10)

Arguments

data
a table with columns id (character), code, and optionally date (Date)

codedata
additional data with columns including case id (character), code and an optional date (Date) for each code. An optional column condition might distinguish codes/dates with certain characteristics (see example).

...
arguments passed between methods

id, code, date, code_date
column names with case id (character from data and codedata), code (from data) and optional date (Date from data) and code_date (Date from codedata).

days
numeric vector of length two with lower and upper bound for range of relevant days relative to date. See "Relevant period".

alnum
Should codes be cleaned from all non alphanumeric characters?

.copy
Should the object be copied internally by data.table::copy()? NA (by default) means that objects smaller than 1 GB are copied. If the size is larger, the argument must be set explicitly. Set TRUE to make copies regardless of object size. This is recommended if enough RAM is available. If set to FALSE, calculations might be carried out but the object will be changed by reference.

x
object (potentially of large size)

n
number of rows to preview as tibble. The output is technically a data.table::data.table, which might be an unusual format to look at. Use n = NULL to print the object as is.

Value

Object of class codified (inheriting from data.table::data.table). Essentially data with additional columns: code, code_date: left joined from codedata or NA if no match within period. in_period: Boolean indicator if the case had at least one code within the specified period.

The output has one row for each combination of "id" from data and "code" from codedata. Rows from data might be repeated accordingly.

Relevant period

Some examples for argument days:

c(-365, -1): window of one year prior to the date column of data. Useful for patient co-morbidity.

c(1, 30): window of 30 days after date. Useful for adverse events after a surgical procedure.

c(-Inf, Inf): no limitation on non missing dates.

NULL: no time limitation at all.

See also

Examples

codify(ex_people, ex_icd10, id = "name", code = "icd10",
  date = "surgery", code_date = "admission", days = c(-365, 0))

#> 
#> The printed data is of class: codified, data.table, data.frame. It has 378 rows.
#> It is here previewed as a tibble
#> Use `print(x, n = NULL)` to print as is (or use `n` to specify the number of rows to preview)!
#> 
#> 
#> # A tibble: 10 x 6
#>    name                surgery    admission  icd10 hdia  in_period
#>    <chr>               <date>     <date>     <chr> <lgl> <lgl>    
#>  1 Archer, Leon Hunter 2020-05-25 2019-12-06 E012  FALSE TRUE     
#>  2 Archer, Leon Hunter 2020-05-25 2019-12-28 V8698 FALSE TRUE     
#>  3 Archer, Leon Hunter 2020-05-25 2020-01-17 X4128 FALSE TRUE     
#>  4 Archer, Leon Hunter 2020-05-25 2020-01-18 X3403 FALSE TRUE     
#>  5 Archer, Leon Hunter 2020-05-25 2020-02-03 Z752  FALSE TRUE     
#>  6 Archer, Leon Hunter 2020-05-25 2020-02-11 V7413 FALSE TRUE     
#>  7 Awtrey, Antonio     2020-09-20 2020-01-14 X3322 FALSE TRUE     
#>  8 Awtrey, Antonio     2020-09-20 2020-04-05 Y1614 FALSE TRUE     
#>  9 Awtrey, Antonio     2020-09-20 2020-04-08 X7564 FALSE TRUE     
#> 10 Awtrey, Antonio     2020-09-20 2020-05-26 X6542 FALSE TRUE