codify elements

codify elements

codify(x, from, id = "id", date = NULL, days = NULL, .copy = NA)

Arguments

x
data.frame with at least two columns, one with case (patient) identification (column name specified by argument id) and one with a date of interest (column name specified by argument date)

from
object with code data for which is.codedata is TRUE

id
name of column in x containing case (patient) identification

date
name of column in x with possible date of interest (NULL by default but must be specified if argument 'days' is used).

days
numeric vector of length two with lower and upper bound of range of days relative to date for which codes from from are relevant. (For example c(-365, -1) implies a time window of one year prior to date, which might be useful for calculating comorbidity indices, while c(1, 30) gives a window of 30 days after date, which might be used for calculating adverse events after a surgical procedure.) c(-Inf, Inf) means no limitation on non missing dates. NULL means no time limitation at all.

.copy
Should the object be copied internally by copy? NA (by default) means that objects smaller than 1 Gb are copied. If the size is larger, the argument must be set explicitly. Set TRUE to make copies regardless of object size. This is recomended if enough RAM is available. If set to FALSE, calculations might be carried out but the object will be changed by reference.

Value

Data frame (data.table) with columns corresponding to x and additional columns matched from from:

code: code as matched from from (NA if no match within period)

code_date: corresponding date for which the code was valid (NA if no match within period)

in_period: boolean indicator if the unit (patient) had at least one code within the specified period

The output has at one row for each combination of "id" and code. Note that other columns of x might be repeated accordingly.

See also

Examples

codify(ex_people, ex_icd10, id = "name", date = "surgery", days = c(-365, 0))

#>                  name       date  code_date  code  hdia in_period
#>   1:    Adams, Jaimin 2019-06-14 2018-08-16  E835 FALSE      TRUE
#>   2:    Adams, Jaimin 2019-06-14 2018-10-03  N179 FALSE      TRUE
#>   3:    Adams, Jaimin 2019-06-14 2018-12-01  M713 FALSE      TRUE
#>   4:    Adams, Jaimin 2019-06-14 2018-12-07 M861F FALSE      TRUE
#>   5:    Adams, Jaimin 2019-06-14 2019-01-29  T323 FALSE      TRUE
#>  ---                                                             
#> 354: el-Zadeh,  Muzna 2018-12-28 2018-06-11  Q753 FALSE      TRUE
#> 355: el-Zadeh,  Muzna 2018-12-28 2018-11-26 M000G FALSE      TRUE
#> 356: el-Zadeh,  Muzna 2018-12-28 2018-12-03  T589 FALSE      TRUE
#> 357: el-Zadeh,  Muzna 2018-12-28 2018-12-10  C946 FALSE      TRUE
#> 358: el-Zadeh,  Muzna 2018-12-28 2018-12-19  J011 FALSE      TRUE