Calculate index based on classification scheme

(Weighted) sum of all identified classes for each case. Index weights for subordinate hierarchical classes (as identified by attr(cc, "hierarchy")) are excluded in presence of superior classes if index specified with argument index.

index(classified, ...)

# S3 method for data.frame
index(classified, ...)

# S3 method for matrix
index(classified, index = NULL, cc = NULL, ...)

Arguments

classified
output from classify()

...
used internally

index
name of column with 'weights' from corresponding classcodes object. Can be NULL if the index is just a count of relevant classes.

cc
classcodes object. Can be NULL if information already present as attribute of classified (which is often the case) and/or if index calculated without weights.

Value

Named numeric index vector with names corresponding to rownames(x)

See also

Examples

# Prepare some codified data with ICD-10 codes during 1 year (365 days)
# before surgery
x <-
  codify(
    ex_people,
    ex_icd10,
    id        = "name",
    code      = "icd10",
    date      = "surgery",
    days      = c(-365, 0),
    code_date = "admission"
  )

# Classify those patients by the Charlson and Elixhasuer comorbidity indices
cl <- classify(x, "charlson")

#> Warning: 'classify()' does not preserve row order ('categorize()' does!)

#> Classification based on: icd10

# Calculate (weighted) index values
index(cl)                  # Un-weighted sum/no of conditions for each patient

#> index calculated as number of relevant categories

#>        Connelly, Monta   Garcia Hyett, Brenda  Martinez Botello, Luz 
#>                      0                      0                      0 
#>      Marton, Gabriella         Mitchell, John           Nelson, Cory 
#>                      0                      0                      0 
#>          Oh, Stephanie          Terry, Calvin     Williams, Nickolas 
#>                      0                      0                      0 
#>        Young, Mitchell      Benitez, Veronica          Deleon, Isaac 
#>                      0                     NA                     NA 
#>           Ebat, Kabita         Gerber, Alexis           Hill, Audrey 
#>                     NA                     NA                     NA 
#>        Howling, Keisha       Kauffmann, Kelia          Lopez, Daniel 
#>                     NA                     NA                     NA 
#>        Nguyen, Camille        Pracht, Derrick   Santistevan, Charlie 
#>                     NA                     NA                     NA 
#>         Schmer, Taylor   Sommerville, Dominic           Winn, Robert 
#>                     NA                     NA                     NA 
#>      el-Kaba, Ghuzaila     al-Akhter, Marzooq    al-Hossain, Zuraara 
#>                     NA                      0                      1 
#>   al-Koroma, Humaidaan       al-Vohra, Jamaal    Archer, Leon Hunter 
#>                      0                      1                      0 
#>        Awtrey, Antonio     Bammesberger, Jozi         Banks, Silbret 
#>                      0                      0                      0 
#>       Begaye, Sullivan            Bhatt, Yuko        Boucher, Teresa 
#>                      0                      0                      1 
#>         Carnes, Shelby        Chapa, Nicholas           Chen, Trevor 
#>                      1                      0                      0 
#>           Chin, Nelson Christensen, Johnathon        Cleveland, Mark 
#>                      0                      1                      0 
#>       Cornelius, Kelly       el-Abdoo, Minnah        el-Mirza, Ismad 
#>                      0                      0                      0 
#>        el-Sani, Jawaad      el-Shahin, Hasnaa                Fell, J 
#>                      0                      0                      0 
#>       Fuentes, Roberto           Garza, Kenia        Glover, Phillip 
#>                      1                      0                      0 
#>        Graves, Acineth       Hammond, Breanna     Hernandez, Anthony 
#>                      0                      0                      0 
#>         Howard, Darius           Irick, Kevin         Jaime, Cynthia 
#>                      0                      0                      0 
#>    Jefferson, Deontrae       Johnson, William             Kang, Xuan 
#>                      0                      0                      0 
#>          Lacsina, Ryan            Lam, Hannah          Le, Christina 
#>                      0                      0                      1 
#>            Le, Soraiya           Lee, Jaeyeon           Lehi, Marilu 
#>                      0                      0                      0 
#>         Lening, Hannah            Lor, Vuthdy          Mahat, Sydney 
#>                      1                      0                      0 
#>        Mandakh, Joseph       Martinez, Magler        Mcleod, Desmond 
#>                      0                      1                      0 
#>          Meier, Hayden          Murphy, Kevin            Ngu, Brenda 
#>                      1                      0                      0 
#>     Nikolaeff, Brandon        Norton, Charles       Packard, Kaitlin 
#>                      0                      0                      0 
#>   Perales, Christopher         Rico, Germaine       Robinson, Markel 
#>                      0                      0                      0 
#>         Rodri, Tiffany          Rubin, Miriah      Shuemaker, Lauren 
#>                      0                      0                      0 
#>       Simpson, Kenneth          Slater, Trina       Steinke, Richard 
#>                      0                      0                      0 
#>          Surage, Sydne        Swimmer, Apryll        Thumma, Phillip 
#>                      0                      0                      0 
#>          Trine, Alyssa        Trinh, Schuyler       Trujillo, Olivia 
#>                      0                      0                      2 
#>      Trujillo, Yanelly         Tveten, Austin      Weddington, Omega 
#>                      0                      0                      0 
#>        Williams, Ellen          Ybarra, Maria          Yost, Rebecca 
#>                      0                      0                      0 
#>            Zhen, Cassa 
#>                      0 

index(cl, "quan_original") # Weighted index (Quan et al. 2005; see `?charlson`)

#>        Connelly, Monta   Garcia Hyett, Brenda  Martinez Botello, Luz 
#>                      0                      0                      0 
#>      Marton, Gabriella         Mitchell, John           Nelson, Cory 
#>                      0                      0                      0 
#>          Oh, Stephanie          Terry, Calvin     Williams, Nickolas 
#>                      0                      0                      0 
#>        Young, Mitchell      Benitez, Veronica          Deleon, Isaac 
#>                      0                     NA                     NA 
#>           Ebat, Kabita         Gerber, Alexis           Hill, Audrey 
#>                     NA                     NA                     NA 
#>        Howling, Keisha       Kauffmann, Kelia          Lopez, Daniel 
#>                     NA                     NA                     NA 
#>        Nguyen, Camille        Pracht, Derrick   Santistevan, Charlie 
#>                     NA                     NA                     NA 
#>         Schmer, Taylor   Sommerville, Dominic           Winn, Robert 
#>                     NA                     NA                     NA 
#>      el-Kaba, Ghuzaila     al-Akhter, Marzooq    al-Hossain, Zuraara 
#>                     NA                      0                      6 
#>   al-Koroma, Humaidaan       al-Vohra, Jamaal    Archer, Leon Hunter 
#>                      0                      2                      0 
#>        Awtrey, Antonio     Bammesberger, Jozi         Banks, Silbret 
#>                      0                      0                      0 
#>       Begaye, Sullivan            Bhatt, Yuko        Boucher, Teresa 
#>                      0                      0                      1 
#>         Carnes, Shelby        Chapa, Nicholas           Chen, Trevor 
#>                      1                      0                      0 
#>           Chin, Nelson Christensen, Johnathon        Cleveland, Mark 
#>                      0                      1                      0 
#>       Cornelius, Kelly       el-Abdoo, Minnah        el-Mirza, Ismad 
#>                      0                      0                      0 
#>        el-Sani, Jawaad      el-Shahin, Hasnaa                Fell, J 
#>                      0                      0                      0 
#>       Fuentes, Roberto           Garza, Kenia        Glover, Phillip 
#>                      2                      0                      0 
#>        Graves, Acineth       Hammond, Breanna     Hernandez, Anthony 
#>                      0                      0                      0 
#>         Howard, Darius           Irick, Kevin         Jaime, Cynthia 
#>                      0                      0                      0 
#>    Jefferson, Deontrae       Johnson, William             Kang, Xuan 
#>                      0                      0                      0 
#>          Lacsina, Ryan            Lam, Hannah          Le, Christina 
#>                      0                      0                      2 
#>            Le, Soraiya           Lee, Jaeyeon           Lehi, Marilu 
#>                      0                      0                      0 
#>         Lening, Hannah            Lor, Vuthdy          Mahat, Sydney 
#>                      2                      0                      0 
#>        Mandakh, Joseph       Martinez, Magler        Mcleod, Desmond 
#>                      0                      1                      0 
#>          Meier, Hayden          Murphy, Kevin            Ngu, Brenda 
#>                      1                      0                      0 
#>     Nikolaeff, Brandon        Norton, Charles       Packard, Kaitlin 
#>                      0                      0                      0 
#>   Perales, Christopher         Rico, Germaine       Robinson, Markel 
#>                      0                      0                      0 
#>         Rodri, Tiffany          Rubin, Miriah      Shuemaker, Lauren 
#>                      0                      0                      0 
#>       Simpson, Kenneth          Slater, Trina       Steinke, Richard 
#>                      0                      0                      0 
#>          Surage, Sydne        Swimmer, Apryll        Thumma, Phillip 
#>                      0                      0                      0 
#>          Trine, Alyssa        Trinh, Schuyler       Trujillo, Olivia 
#>                      0                      0                      3 
#>      Trujillo, Yanelly         Tveten, Austin      Weddington, Omega 
#>                      0                      0                      0 
#>        Williams, Ellen          Ybarra, Maria          Yost, Rebecca 
#>                      0                      0                      0 
#>            Zhen, Cassa 
#>                      0 

index(cl, "quan_updated")  # Weighted index (Quan et al. 2011; see `?charlson`)

#>        Connelly, Monta   Garcia Hyett, Brenda  Martinez Botello, Luz 
#>                      0                      0                      0 
#>      Marton, Gabriella         Mitchell, John           Nelson, Cory 
#>                      0                      0                      0 
#>          Oh, Stephanie          Terry, Calvin     Williams, Nickolas 
#>                      0                      0                      0 
#>        Young, Mitchell      Benitez, Veronica          Deleon, Isaac 
#>                      0                     NA                     NA 
#>           Ebat, Kabita         Gerber, Alexis           Hill, Audrey 
#>                     NA                     NA                     NA 
#>        Howling, Keisha       Kauffmann, Kelia          Lopez, Daniel 
#>                     NA                     NA                     NA 
#>        Nguyen, Camille        Pracht, Derrick   Santistevan, Charlie 
#>                     NA                     NA                     NA 
#>         Schmer, Taylor   Sommerville, Dominic           Winn, Robert 
#>                     NA                     NA                     NA 
#>      el-Kaba, Ghuzaila     al-Akhter, Marzooq    al-Hossain, Zuraara 
#>                     NA                      0                      6 
#>   al-Koroma, Humaidaan       al-Vohra, Jamaal    Archer, Leon Hunter 
#>                      0                      2                      0 
#>        Awtrey, Antonio     Bammesberger, Jozi         Banks, Silbret 
#>                      0                      0                      0 
#>       Begaye, Sullivan            Bhatt, Yuko        Boucher, Teresa 
#>                      0                      0                      0 
#>         Carnes, Shelby        Chapa, Nicholas           Chen, Trevor 
#>                      1                      0                      0 
#>           Chin, Nelson Christensen, Johnathon        Cleveland, Mark 
#>                      0                      0                      0 
#>       Cornelius, Kelly       el-Abdoo, Minnah        el-Mirza, Ismad 
#>                      0                      0                      0 
#>        el-Sani, Jawaad      el-Shahin, Hasnaa                Fell, J 
#>                      0                      0                      0 
#>       Fuentes, Roberto           Garza, Kenia        Glover, Phillip 
#>                      2                      0                      0 
#>        Graves, Acineth       Hammond, Breanna     Hernandez, Anthony 
#>                      0                      0                      0 
#>         Howard, Darius           Irick, Kevin         Jaime, Cynthia 
#>                      0                      0                      0 
#>    Jefferson, Deontrae       Johnson, William             Kang, Xuan 
#>                      0                      0                      0 
#>          Lacsina, Ryan            Lam, Hannah          Le, Christina 
#>                      0                      0                      2 
#>            Le, Soraiya           Lee, Jaeyeon           Lehi, Marilu 
#>                      0                      0                      0 
#>         Lening, Hannah            Lor, Vuthdy          Mahat, Sydney 
#>                      2                      0                      0 
#>        Mandakh, Joseph       Martinez, Magler        Mcleod, Desmond 
#>                      0                      0                      0 
#>          Meier, Hayden          Murphy, Kevin            Ngu, Brenda 
#>                      0                      0                      0 
#>     Nikolaeff, Brandon        Norton, Charles       Packard, Kaitlin 
#>                      0                      0                      0 
#>   Perales, Christopher         Rico, Germaine       Robinson, Markel 
#>                      0                      0                      0 
#>         Rodri, Tiffany          Rubin, Miriah      Shuemaker, Lauren 
#>                      0                      0                      0 
#>       Simpson, Kenneth          Slater, Trina       Steinke, Richard 
#>                      0                      0                      0 
#>          Surage, Sydne        Swimmer, Apryll        Thumma, Phillip 
#>                      0                      0                      0 
#>          Trine, Alyssa        Trinh, Schuyler       Trujillo, Olivia 
#>                      0                      0                      1 
#>      Trujillo, Yanelly         Tveten, Austin      Weddington, Omega 
#>                      0                      0                      0 
#>        Williams, Ellen          Ybarra, Maria          Yost, Rebecca 
#>                      0                      0                      0 
#>            Zhen, Cassa 
#>                      0 

# Tabulate index for all patients.
# As expected, most patients are healthy and have index = 0/NA,
# where NA indicates no recorded hospital visits
# found in `ex_icd10` during codification.
# In practice, those patients might be assumed to have 0 comorbidity as well.
table(index(cl, "quan_original"), useNA = "always")

#> 
#>    0    1    2    3    6 <NA> 
#>   74    5    4    1    1   15 

# If `cl` is a matrix without additional attributes (as imposed by `codify()`)
# an explicit classcodes object must be specified by the `cc` argument
cl2 <- as.matrix(cl)
index(cl2, cc = "charlson")

#> index calculated as number of relevant categories

#>        Connelly, Monta   Garcia Hyett, Brenda  Martinez Botello, Luz 
#>                      0                      0                      0 
#>      Marton, Gabriella         Mitchell, John           Nelson, Cory 
#>                      0                      0                      0 
#>          Oh, Stephanie          Terry, Calvin     Williams, Nickolas 
#>                      0                      0                      0 
#>        Young, Mitchell      Benitez, Veronica          Deleon, Isaac 
#>                      0                     NA                     NA 
#>           Ebat, Kabita         Gerber, Alexis           Hill, Audrey 
#>                     NA                     NA                     NA 
#>        Howling, Keisha       Kauffmann, Kelia          Lopez, Daniel 
#>                     NA                     NA                     NA 
#>        Nguyen, Camille        Pracht, Derrick   Santistevan, Charlie 
#>                     NA                     NA                     NA 
#>         Schmer, Taylor   Sommerville, Dominic           Winn, Robert 
#>                     NA                     NA                     NA 
#>      el-Kaba, Ghuzaila     al-Akhter, Marzooq    al-Hossain, Zuraara 
#>                     NA                      0                      1 
#>   al-Koroma, Humaidaan       al-Vohra, Jamaal    Archer, Leon Hunter 
#>                      0                      1                      0 
#>        Awtrey, Antonio     Bammesberger, Jozi         Banks, Silbret 
#>                      0                      0                      0 
#>       Begaye, Sullivan            Bhatt, Yuko        Boucher, Teresa 
#>                      0                      0                      1 
#>         Carnes, Shelby        Chapa, Nicholas           Chen, Trevor 
#>                      1                      0                      0 
#>           Chin, Nelson Christensen, Johnathon        Cleveland, Mark 
#>                      0                      1                      0 
#>       Cornelius, Kelly       el-Abdoo, Minnah        el-Mirza, Ismad 
#>                      0                      0                      0 
#>        el-Sani, Jawaad      el-Shahin, Hasnaa                Fell, J 
#>                      0                      0                      0 
#>       Fuentes, Roberto           Garza, Kenia        Glover, Phillip 
#>                      1                      0                      0 
#>        Graves, Acineth       Hammond, Breanna     Hernandez, Anthony 
#>                      0                      0                      0 
#>         Howard, Darius           Irick, Kevin         Jaime, Cynthia 
#>                      0                      0                      0 
#>    Jefferson, Deontrae       Johnson, William             Kang, Xuan 
#>                      0                      0                      0 
#>          Lacsina, Ryan            Lam, Hannah          Le, Christina 
#>                      0                      0                      1 
#>            Le, Soraiya           Lee, Jaeyeon           Lehi, Marilu 
#>                      0                      0                      0 
#>         Lening, Hannah            Lor, Vuthdy          Mahat, Sydney 
#>                      1                      0                      0 
#>        Mandakh, Joseph       Martinez, Magler        Mcleod, Desmond 
#>                      0                      1                      0 
#>          Meier, Hayden          Murphy, Kevin            Ngu, Brenda 
#>                      1                      0                      0 
#>     Nikolaeff, Brandon        Norton, Charles       Packard, Kaitlin 
#>                      0                      0                      0 
#>   Perales, Christopher         Rico, Germaine       Robinson, Markel 
#>                      0                      0                      0 
#>         Rodri, Tiffany          Rubin, Miriah      Shuemaker, Lauren 
#>                      0                      0                      0 
#>       Simpson, Kenneth          Slater, Trina       Steinke, Richard 
#>                      0                      0                      0 
#>          Surage, Sydne        Swimmer, Apryll        Thumma, Phillip 
#>                      0                      0                      0 
#>          Trine, Alyssa        Trinh, Schuyler       Trujillo, Olivia 
#>                      0                      0                      2 
#>      Trujillo, Yanelly         Tveten, Austin      Weddington, Omega 
#>                      0                      0                      0 
#>        Williams, Ellen          Ybarra, Maria          Yost, Rebecca 
#>                      0                      0                      0 
#>            Zhen, Cassa 
#>                      0