Calculate index based on classification scheme

Calculate index based on classification scheme

index(x, ...)

# S3 method for data.frame
index(x, ...)

# S3 method for matrix
index(x, by = NULL, from = NULL, ...)

Arguments

x
output from classify (or possibly from as.data.frame.classified)

...
used internally

by
name of column with 'weights' from corresponding classcodes object. Can be omitted if the index is just the count of relevant classes.

from
classcodes object. Can be omitted if information already present as attribute of x (which is often the case) and/or if index calculated without weights.

Value

Named numeric index vector with names corresponding to rownames(x)

See also

Examples

# Calculate Elixhauser comorbidity index
cl <- classify(c("C80", "I20"), "elix_icd10")
index(cl)

#> index calculated as number of relevant categories

#> C80 I20 
#>   1   0 

# Calculate Charlson-index using original weights
cl <- classify(c("C80", "I21"), "charlson_icd10")
index(cl, "quan_original")

#> C80 I21 
#>   6   1 

# Calculate Charlson-index using updated weights
cl <- classify(c("C80", "I21"), "charlson_icd10")
index(cl, "quan_updated")

#> C80 I21 
#>   6   0 

# Find patients with adverse events after hip surgery
co <- codify(ex_people, ex_icd10, id = "name",
    date = "surgery", days = c(-365, 0))
cl <- classify(co, "hip_adverse_events_icd10_old")
index(cl)

#> index calculated as number of relevant categories

#>           Alexis, Dayveon            Bauder, Taylor          Beasley, Michael 
#>                         0                         0                         0 
#>         Deshazor, Kordell         Henricks, Donovan       Marquez, Jacqueline 
#>                         0                         0                         0 
#>           Martinez, David         Mccarthy, Gabriel             Simpson, King 
#>                         0                         0                         0 
#>           Gallegos, Diana          Perez Jr, Hannah             Reyes, Sylvia 
#>                        NA                        NA                        NA 
#>              Rusnak, Kyle            Ward, Harrison      al-Kassem, Humaidaan 
#>                        NA                        NA                        NA 
#>          el-Nagi, Waddaah            Abzari, Joseph             al-Abid, Asad 
#>                        NA                         0                         0 
#>            al-Ahsan, Umar       al-Eid, Abdus Samad         al-Mansur, Kawkab 
#>                         0                         0                         0 
#>         Anderson, Laquesh        Armijo Jr, Kalynne           Babcock, Landon 
#>                         0                         0                         0 
#>             Barela, Marco          Barger, Willanna              Barnes, Amos 
#>                         0                         0                         0 
#>           Beaver, Tristin           Bille, Cheyenne           Boffill, Jordan 
#>                         0                         0                         1 
#>            Bradshaw, Noah          Breshears, Kayla              Bruce, Bryan 
#>                         0                         0                         0 
#>              Bruno, Jorge          Burton, Danyalle           Charlie, Mareno 
#>                         0                         0                         0 
#> Dominguez Enriquez, David      Edmondson, Christian          el-Assad, Jubair 
#>                         0                         0                         0 
#>          el-Bina, Muntaha          el-Kazi, Najeeba        el-Masood, Nasreen 
#>                         0                         0                         0 
#>            el-Riaz, Aadam            el-Wakim, Nuha        el-Zakaria, Saajid 
#>                         0                         0                         0 
#>              Elder, Kevyn          Ghimire, Chelsea             Hall, Samorra 
#>                         0                         0                         0 
#>            Hayes, Kylihah            Headden, Devon                  Her, Lia 
#>                         0                         0                         0 
#>             Hicks, Camren             Hoang, Elijah     Jiles-Wright, Bralynd 
#>                         0                         0                         0 
#>               Jung, Derek              Kane, Chapin                Kim, Alisa 
#>                         0                         0                         0 
#>                King, Hali          Laulu, Kajtshiab             Lopez, Chiara 
#>                         0                         0                         0 
#>         Maestas, Lilibeth                Mamet, A D          Martin, Elizario 
#>                         0                         0                         0 
#>         Martinez, Jessica            Martinez, Juan          Martinez, Nelson 
#>                         0                         0                         0 
#>         Mcginnis, Adunola          Mcknight, Jaymes           Miller, Darshay 
#>                         0                         0                         0 
#>          Moktader, Sameer            Mosely, Jereek                 Nam, John 
#>                         0                         0                         0 
#>           Nelson, Phillip            Okoye, Vanessa             Parker, Pharo 
#>                         0                         0                         0 
#>           Perez, Kimberly                Pham, Kyle          Pronteau, Colton 
#>                         0                         0                         0 
#>           Rodriguez, Alma    Rowley Booneel, Hailee        Sanburg, Cassandra 
#>                         0                         0                         0 
#>        Sanchez, Dominique             Santee, Holly              Schell, Jack 
#>                         0                         0                         0 
#>         Schubert, Patrick               Smith, John             Smith, Ruasha 
#>                         0                         0                         0 
#>              Spoerl, John            Stephan, Scott           Suriwong, Sarah 
#>                         0                         0                         1 
#>         Thompson, Stephen          Thwaites, Travis       Todacheene, Erminio 
#>                         0                         0                         0 
#>         Tomlinson, Amanda    Tuitele-Britton, Janet          Valencia, Johnny 
#>                         0                         0                         0 
#>         Vallie, Mitchelle           Vigo, Shi Hyung          Wilkerson, Teesa 
#>                         0                         0                         0 
#>              Wood, Elijah 
#>                         0