Calculate index based on classification scheme

(Weighted) sum of all identified classes for each case. Index weights for subordinate hierarchical classes (as identified by attr(cc, "hierarchy")) are excluded in presence of superior classes if index specified with argument index.

index(classified, ...)

# S3 method for data.frame
index(classified, ...)

# S3 method for matrix
index(classified, index = NULL, cc = NULL, ...)

Arguments

classified
output from classify()

...
used internally

index
name of column with 'weights' from corresponding classcodes object. Can be NULL if the index is just a count of relevant classes.

cc
classcodes object. Can be NULL if information already present as attribute of classified (which is often the case) and/or if index calculated without weights.

Value

Named numeric index vector with names corresponding to rownames(x)

See also

Examples

# Calculate Elixhauser comorbidity index
cl <- classify(c("C80", "I20"), "elixhauser")

#> Classification based on: icd10

index(cl)

#> index calculated as number of relevant categories

#> C80 I20 
#>   1   0 

# Calculate Charlson-index using original weights
cl <- classify(c("C80", "I21"), "charlson")

#> Classification based on: icd10

index(cl, "quan_original")

#> C80 I21 
#>   6   1 

# Calculate Charlson-index using updated weights
cl <- classify(c("C80", "I21"), "charlson")

#> Classification based on: icd10

index(cl, "quan_updated")

#> C80 I21 
#>   6   0