Printing classified data

Preview first n rows as tibble

# S3 method for classified
print(x, ...)

Arguments

x

output from classify()

...

additional arguments passed to printing method for a tibble. n is the number of rows to preview. Set n = NULL to disable the tibble preview and print the object as is (a matrix).

See also

Examples

# Preview all output
classify(c("C80", "I20", "unvalid_code"), "elixhauser")
#> Classification based on: icd10
#> 
#> The printed data is of class: classified, matrix.
#> It has 3 row(s).
#> It is here previewed as a tibble
#> Use `print(x, n = NULL)` to print as is (or use `n` to specify the number of rows to preview)!
#> 
#> # A tibble: 3 × 31
#>   congestive h…¹ cardi…² valvu…³ pulmo…⁴ perip…⁵ hyper…⁶ hyper…⁷ paral…⁸ other…⁹
#>   <lgl>          <lgl>   <lgl>   <lgl>   <lgl>   <lgl>   <lgl>   <lgl>   <lgl>  
#> 1 FALSE          FALSE   FALSE   FALSE   FALSE   FALSE   FALSE   FALSE   FALSE  
#> 2 FALSE          FALSE   FALSE   FALSE   FALSE   FALSE   FALSE   FALSE   FALSE  
#> 3 FALSE          FALSE   FALSE   FALSE   FALSE   FALSE   FALSE   FALSE   FALSE  
#> # … with 22 more variables: `chronic pulmonary disease` <lgl>,
#> #   `diabetes uncomplicated` <lgl>, `diabetes complicated` <lgl>,
#> #   hypothyroidism <lgl>, `renal failure` <lgl>, `liver disease` <lgl>,
#> #   `peptic ulcer disease` <lgl>, `AIDS/HIV` <lgl>, lymphoma <lgl>,
#> #   `metastatic cancer` <lgl>, `solid tumor` <lgl>,
#> #   `rheumatoid arthritis` <lgl>, coagulopathy <lgl>, obesity <lgl>,
#> #   `weight loss` <lgl>, `fluid electrolyte disorders` <lgl>, …

# Preview only the first row
print(classify(c("C80", "I20", "unvalid_code"), "elixhauser"), n = 1)
#> Classification based on: icd10
#> 
#> The printed data is of class: classified, matrix.
#> It has 3 row(s).
#> It is here previewed as a tibble
#> Use `print(x, n = NULL)` to print as is (or use `n` to specify the number of rows to preview)!
#> 
#> # A tibble: 1 × 31
#>   congestive h…¹ cardi…² valvu…³ pulmo…⁴ perip…⁵ hyper…⁶ hyper…⁷ paral…⁸ other…⁹
#>   <lgl>          <lgl>   <lgl>   <lgl>   <lgl>   <lgl>   <lgl>   <lgl>   <lgl>  
#> 1 FALSE          FALSE   FALSE   FALSE   FALSE   FALSE   FALSE   FALSE   FALSE  
#> # … with 22 more variables: `chronic pulmonary disease` <lgl>,
#> #   `diabetes uncomplicated` <lgl>, `diabetes complicated` <lgl>,
#> #   hypothyroidism <lgl>, `renal failure` <lgl>, `liver disease` <lgl>,
#> #   `peptic ulcer disease` <lgl>, `AIDS/HIV` <lgl>, lymphoma <lgl>,
#> #   `metastatic cancer` <lgl>, `solid tumor` <lgl>,
#> #   `rheumatoid arthritis` <lgl>, coagulopathy <lgl>, obesity <lgl>,
#> #   `weight loss` <lgl>, `fluid electrolyte disorders` <lgl>, …

# Print object as is (matrix)
print(classify(c("C80", "I20", "unvalid_code"), "elixhauser"), n = NULL)
#> Classification based on: icd10
#> 
#> The printed data is of class: classified, matrix.
#> It has 3 row(s).
#> It is here previewed as a tibble
#> Use `print(x, n = NULL)` to print as is (or use `n` to specify the number of rows to preview)!
#> 
#> # A tibble: 3 × 31
#>   congestive h…¹ cardi…² valvu…³ pulmo…⁴ perip…⁵ hyper…⁶ hyper…⁷ paral…⁸ other…⁹
#>   <lgl>          <lgl>   <lgl>   <lgl>   <lgl>   <lgl>   <lgl>   <lgl>   <lgl>  
#> 1 FALSE          FALSE   FALSE   FALSE   FALSE   FALSE   FALSE   FALSE   FALSE  
#> 2 FALSE          FALSE   FALSE   FALSE   FALSE   FALSE   FALSE   FALSE   FALSE  
#> 3 FALSE          FALSE   FALSE   FALSE   FALSE   FALSE   FALSE   FALSE   FALSE  
#> # … with 22 more variables: `chronic pulmonary disease` <lgl>,
#> #   `diabetes uncomplicated` <lgl>, `diabetes complicated` <lgl>,
#> #   hypothyroidism <lgl>, `renal failure` <lgl>, `liver disease` <lgl>,
#> #   `peptic ulcer disease` <lgl>, `AIDS/HIV` <lgl>, lymphoma <lgl>,
#> #   `metastatic cancer` <lgl>, `solid tumor` <lgl>,
#> #   `rheumatoid arthritis` <lgl>, coagulopathy <lgl>, obesity <lgl>,
#> #   `weight loss` <lgl>, `fluid electrolyte disorders` <lgl>, …