Classcodes for RxRisk V based on ATC codes

classcodes object to use with classify. Provided mostly as proof-of-concept. Codes have not been externaly validated and desired implemantation might differ over time and by country.

rxriskv

Format

Data frames with 39 rows and 6 variables:

group

medical condition

regex

ATC codes from table 1 in Pratt et al. 2018 (ignoring PBS item codes and extra conditions).

regex_unknown

Source for this classification is unknown!

regex_garland

Based on regex_unknown but modified by Anne Garland to resemble medical use in Sweden 2016. Unpublished!

regex_caughey

From appendix 1 in Caughey et al. 2010

pratt

Mortality weights from table 1 in Pratt et al. 2018

References

Caughey GE, Roughead EE, Vitry AI, McDermott RA, Shakib S, Gilbert AL. Comorbidity in the elderly with diabetes: Identification of areas of potential treatment conflicts. Diabetes Res Clin Pract 2010;87:385–93. doi:10.1016/j.diabres.2009.10.019.

Pratt NL, Kerr M, Barratt JD, Kemp-Casey A, Kalisch Ellett LM, Ramsay E, et al. The validity of the Rx-Risk Comorbidity Index using medicines mapped to the Anatomical Therapeutic Chemical (ATC) Classification System. BMJ Open 2018;8. doi:10.1136/bmjopen-2017-021122.

See also