Summarizing a classcodes object

Classification schemes are formalized by regular expressions within the classcodes objects. These are computationally effective but sometimes hard to grasp. Use this function to present all individual codes identified for each group in the classification.

# S3 method for classcodes
summary(object, coding, ..., cc_args = list())

Arguments

object
classcodes object

coding
either a vector with codes from the original classification, or a name (character vector of length one) keyvalue object from package "decoder" (for example "icd10cm" or "atc")

...
ignored

cc_args
List of named arguments passed to set_classcodes()

Value

List (invisible) with objects:

object: input object

summary: a data frame with columns:

group: Groups identified by object.

n: The number of codes to be recognized for each group.

codes: Individual codes within each group.

coding: input coding

See also

Examples

summary(elixhauser, coding = "icd10cm")

#> Classification based on: icd10

#> Indices: sum_all, sum_all_ahrq, walraven, sid29, sid30, ahrq_mort, ahrq_readm 
#> 
#> Recognized codes per group:
#> 
#> # A tibble: 31 x 3
#>    group                   n codes                                              
#>    <chr>               <int> <chr>                                              
#>  1 AIDS/HIV                1 B20                                                
#>  2 alcohol abuse         131 E52, F1010, F1011, F10120, F10121, F10129, F1014, ~
#>  3 blood loss anemia       1 D500                                               
#>  4 cardiac arrhythmias    73 I441, I442, I4430, I4439, I456, I459, I470, I471, ~
#>  5 chronic pulmonary ~    69 I27812, I27822, I27832, I27892, I2792, J40, J410, ~
#>  6 coagulopathy           27 D65, D66, D67, D680, D681, D682, D68311, D68312, D~
#>  7 congestive heart f~    36 I099, I1101, I1301, I1321, I255, I420, I425, I4261~
#>  8 deficiency anemia      17 D508, D509, D510, D511, D512, D513, D518, D519, D5~
#>  9 depression             31 F3130, F3131, F3132, F314, F3152, F320, F321, F322~
#> 10 diabetes complicat~   243 E1021, E1022, E1029, E10311, E10319, E103211, E103~
#> # ... with 21 more rows
#> 
#>  Use function visualize() for a graphical representation.
#>