Summarizing a classcodes object

Classification schemes are formalized by regular expressions within the classcodes objects. These are computationally effective but sometimes hard to grasp. Use this function to present all individual codes identified for each group in the classification.

# S3 method for classcodes
summary(object, regex = "regex", coding, ...)

Arguments

object
classcodes object

regex
name of column with regular expressions to use for classification, either with or without prefix regex_

coding
name (as character) of classification used for coding (for example "icd10" or "atc")

...
ignored

Value

List (invisable) with objects:

object: input object

summary: a data frame with columns:

group: Groups identified by object.

n: The number oif codes to be recognized for each group.

codes: Individual codes within each group.

coding: input coding

See also

Examples

summary(elixhauser, coding = "icd10")

#> 
#> Based on coding: icd10 
#> Indices: sum_all, sum_all_ahrq, walraven, sid29, sid30, ahrq_mort, ahrq_readm 
#> 
#> Recognized codes per group:
#> 
#>  group                          n  
#>  AIDS/HIV                        22
#>  alcohol abuse                   30
#>  blood loss anemia                1
#>  cardiac arrhythmias             45
#>  chronic pulmonary disease       57
#>  coagulopathy                    29
#>  congestive heart failure        19
#>  deficiency anemia               17
#>  depression                      22
#>  diabetes complicated            93
#>  diabetes uncomplicated          33
#>  drug abuse                      82
#>  fluid electrolyte disorders     15
#>  hypertension complicated        13
#>  hypertension uncomplicated       1
#>  hypothyroidism                  21
#>  liver disease                   95
#>  lymphoma                        64
#>  metastatic cancer               29
#>  obesity                          5
#>  other neurological disorders   105
#>  paralysis                       22
#>  peptic ulcer disease             8
#>  peripheral vascular disorder    38
#>  psychoses                       29
#>  pulmonary circulation disorder  10
#>  renal failure                   15
#>  rheumatoid arthritis           107
#>  solid tumor                    421
#>  valvular disease                58
#>  weight loss                     10
#>  codes                                                 
#>  B200, B201, B202, B203, B204, B205, B206, B207, B2 ...
#>  E529, F100, F101, F102, F102A, F102B, F102X, F103, ...
#>  D500                                                  
#>  I441, I441A, I441B, I442, I443, I456, I456A, I456B ...
#>  I2782, I2792, J409, J410, J411, J418, J429, J430,  ...
#>  D659, D669, D679, D680, D681, D682, D683, D684, D6 ...
#>  I099, I1101, I1301, I1321, I255, I420, I425, I4261 ...
#>  D508, D509, D510, D511, D512, D513, D518, D519, D5 ...
#>  F2042, F313, F314, F3152, F320, F321, F322, F323,  ...
#>  E102, E102A, E102B, E102C, E102W, E102X, E103, E10 ...
#>  E100, E100A, E100B, E100C, E100D, E100X, E101, E10 ...
#>  F110, F111, F112, F113, F114, F115, F116, F117, F1 ...
#>  E222, E869, E870, E870A, E870B, E871, E871A, E871B ...
#>  I1102, I119, I1201, I129, I1302, I1311, I1322, I13 ...
#>  I109                                                  
#>  E000, E001, E002, E009, E010, E011, E012, E018, E0 ...
#>  B180, B180A, B180B, B180C, B180D, B180E, B180F, B1 ...
#>  C810, C811, C812, C813, C814, C817, C819, C820, C8 ...
#>  C770, C771, C772, C773, C774, C775, C778, C779, C7 ...
#>  E660, E661, E662, E668, E669                          
#>  G109, G110, G111, G112, G113, G1142, G118, G119, G ...
#>  G041, G1141, G801, G801A, G801B, G801X, G802, G810 ...
#>  K257, K259, K267, K269, K277, K279, K287, K289        
#>  I700, I701, I702, I702A, I702C, I702X, I708, I709, ...
#>  F200, F201, F202, F203, F2041, F205, F206, F208, F ...
#>  I260, I269, I270, I271, I272, I2781, I2791, I280,  ...
#>  I1202, I1312, N181, N182, N183, N184, N185, N189,  ...
#>  L940, L941, L943, M050, M051, M052, M053, M058, M0 ...
#>  C000, C001, C002, C003, C004, C005, C006, C008, C0 ...
#>  A520, I050, I051, I052, I058, I059, I060, I061, I0 ...
#>  E409, E419, E429, E439, E440, E441, E459, E469, R6 ...
#> 
#>  Use function visualize() for a graphical representation.
#>