Visualize classification scheme in web browser

Classes are visualized by their regular expressions in a web browser. The visualization does not give any details on group names, conditions or weights but might be useful both for understanding of a classification scheme in use, and during the creation and debugging of such.

# S3 method for classcodes
visualize(x, group = NULL, show = TRUE, ...)

Arguments

x
classcodes object or name of such object included in the package (see all_classcodes()).

group
names (as character vector) of groups to visualize (all groups if NULL)

show
should a visualization be shown in a web browser. Set to FALSE to just retrieve a URL for later use.

...
Arguments passed on to set_classcodes

regex

name of column with regular expressions to use for classification, either with or without prefix regex_. NULL (default) uses the first classcodes column with prefix "regex". This should be a sensible choice for default classcodes object included in the package.

Value

URL to website with visualization (invisible)

See also

Examples

if (FALSE) {

 # How is depression classified according to Elixhauser?
 visualize("elixhauser", "depression")

 # Compare the two diabetes groups according to Charlson
 visualize("charlson",
   c("diabetes without complication", "diabetes complication"))

 # Is this different from the "Royal College of Surgeons classification?
 # Yes, there is only one group for diabetes
 visualize("charlson",
   c("diabetes without complication", "diabetes complication"),
   regex = "rcs"
 )

 # Show all groups from Charlson
 visualize("charlson")

}

 # Get URL for later visualization
 visualize("hip_ae", show = FALSE)

#> Classification based on: icd10